LST {plotKML}	R Documentation

Time series of MODIS LST images (estimated Land Surface Temperature in degrees C) at 1 km resolution

Description

LST contains a spatial sub-sample (Istra region in Croatia) of 46 time series of MODIS LST images (estimated Land Surface Temperature in degrees C) at 1 km resolution. The "zvalue" slot in the LST object is the dateTime column expressed as yyyy-mm-ddThh:mm:sszzzzzz, where T is the separator between the date and the time, and the time zone is either Z (for UTC) or zzzzzz, which represents ±hh:mm in relation to UTC. The temporal support of these images is 8-days.

Usage

data(LST)

Format

The LST is "RasterBrick" object with the following layers:

LST2008_01_01

8-day MODIS LST mosaick for period 2007-12-29 to 2008-01-04

LST2008_01_09

8-day MODIS LST mosaick for period 2008-01-05 to 2008-01-13

...

...

Note

Time series of 46 day-time and night-time 8-day composite LST images (MOD11A2 product bands 1 and 5) was obtained from the NASA's FTP server (ftp://e4ftl01u.ecs.nasa.gov). The original 8-day composite images were created by patching together images from a period of ±4 days, so that the proportion of clouds can be reduced to a minimum.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl) and Melita Percec Tadic (melita@cirus.dhz.hr)

References

Hengl, T., Heuvelink, G.B.M., Percec Tadic, M., Pebesma, E., (2011) Spatio-temporal prediction of daily temperatures using time-series of MODIS LST images. Theoretical and Applied Climatology.

MODIS products (

Examples


data(LST)
## Not run: 
## Plot temperatures, precipitation and MODIS images in Google Earth:

# to obtain the complete data set for whole Croatia use:
load(url("http://plotkml.r-forge.r-project.org/MODLST.rda"))
## End(Not run)

[Package plotKML version 0.1-2 Index]