RasterBrickTimeSeries-class {plotKML}	R Documentation

A class for a time series of regular grids

Description

A class containing list of rasters, begin, end times and sample points to allow exploration of the values. Objects of this class can be directly visualized in Google Earth by using the plotKML-method.

Slots

variable:

object of class "character"; variable name

sampled:

object of class "SpatialPoints"; one or more points that can be used to visualize temporal changes in the target variable

rasters:

object of class "RasterBrick"; a time-series of raster objects

TimeSpan.begin:

object of class "POSIXct"; begin of sampling for each raster map

TimeSpan.end:

object of class "POSIXct"; end of sampling for each raster map

Methods

plotKML

signature(obj = "RasterBrickTimeSeries"): plots time-series of rasters in Google Earth

Author(s)

Tomislav Hengl

See Also

RasterBrickSimulations-class, plotKML-method

[Package plotKML version 0.5-9 Index]