RasterBrickTimeSeries-class {plotKML}	R Documentation

A class for a regular time series of grids (extends the RasterBrick-class)

Description

A class containing list of rasters, begin, end times and sample points to allow exploration of the values. Objects of this class can be directly visualized in Google Earth by using the plotKML-methods.

Slots

sampled:

object of class "SpatialPoints"; one or more points that can be used to visualize how the values of the feature change through time

rasters:

object of class "RasterBrick"; a time-series of raster objects

TimeSpan.begin:

object of class "POSIXct"; begin of sampling for each raster map

TimeSpan.end:

object of class "POSIXct"; end of sampling for each raster map

Methods

plotKML

signature(obj = "RasterBrickTimeSeries"): plots time-series of rasters in Google Earth

Author(s)

Tomislav Hengl

See Also

RasterBrickSimulations-class, plotKML-methods

Examples

 
# load input data:
data(LST)
## Not run: 
# fit a variogram and generate simulations:
# derive stream networks and save to a list:
# plot the whole project and open in Google Earth:
# plotKML(LST.ls)

## End(Not run)

[Package plotKML version 0.1-4 Index]