SAGA_pal {plotKML}	R Documentation

Colour palettes for numeric variables

Description

SAGA_pal contains 22 colour palettes imported from SAGA GIS (Conrad, 2007). R_pal 12 standard colour palettes used in R to visualize continuous and binary variables. Each colour palette consists of 20 colours in the hexadecimal system. Use display.pal function to plot different sets of palettes.

Usage


data(SAGA_pal)
data(R_pal)

Note

rainbow_75, heat colors, terrain_colors, topo_colors, and bpy_colors are the standard color palettes used in R to visualize numeric/continuous variables. soc_pal, pH_pal, tex_pal, BS_pal and CEC_pal palettes are suitable for visualization of soil variables (soil organic carbon, pH, soil texture fractions, Base Saturation and Cation Exchange Capacity). blue_grey_red palette is recommended for visualization of binary variables (values in the range 0-1), and grey_black is a white-to-black type color palette that contains no white color (hence it will not confuse low values with NA values in the PNG/GIF files).
Possibly the most used palettes for visualization of numeric variables are rev(rainbow(65)[1:48]) and SAGA_pal[[1]] (the SAGA GIS default palette). It is however worth mentioning that in the data visualization literature (and the cartography literature in particular), the rainbow (sometimes also called spectral) color ramp is generally recognized as a bad choice for visualization of sequential/continuous variables (Rogowitz and Treinish, 1998; Borland and Russell, 2007).

Author(s)

SAGA GIS has been created by the SAGA GIS development team (lead by J. Böhner and O. Conrad, from the Institute of Geography, University of Hamburg, Germany). The colour palettes have been exported from SAGA (as ".sprm" SAGA parameter files) and ported to R. All palletes described here were prepared for R by Tomislav Hengl (tom.hengl@wur.nl).

References

Conrad, O., (2007). SAGA — Entwurf, Funktionsumfang und Anwendung eines Systems für Automatisierte Geowissenschaftliche Analysen. Electronic doctoral dissertation, University of Göttingen.

Rogowitz, B.E., Treinish, L.A., (1998, December). Data visualization: the end of the rainbow. Spectrum, IEEE, 35(12):52-59

Borland, D. and Russell, M. T. II, (2007). Rainbow Color Map (Still) Considered Harmful. Computer Graphics and Applications, IEEE, 27(2):14-17.

http://cran.r-project.org/web/packages/RColorBrewer/

http://cran.r-project.org/web/packages/colorspace/

See Also

worldgrids_pal, RColorBrewer::display.brewer.all

Examples


data(SAGA_pal)
data(R_pal)
## Not run: # visualize SAGA GIS palettes:
display.pal(pal=SAGA_pal, sel=c(1,2,7,8,10,11,17,18,19,21,22))
dev.off()
display.pal(R_pal)
names(R_pal)
dev.off()

## End(Not run)

[Package plotKML version 0.2-3 Index]