SpatialMaxEntOutput-class {plotKML}	R Documentation

A class for outputs of analysis produced using the dismo package (MaxEnt)

Description

A class containing input and output data produced by running the maxent (Maximum Entropy) species distribution modeling algorithm. Object of this class can be directly visualized in Google Earth by using the plotKML-method.

Slots

sciname:

object of class "character"; vector of species name compatible with the rgbif package; usually latin "genus" and "species" name

occurrences:

object of class "SpatialPoints"; occurrence-only records

TimeSpan.begin:

object of class "POSIXct"; begin of the sampling period

TimeSpan.end:

object of class "POSIXct"; end of the sampling period

maxent:

object of class "MaxEnt" (species distribution model); produced as an output of the dismo::maxent function

sp.domain:

object of class "Spatial" (ideally "SpatialPolygonsDataFrame" or "SpatialPixelsDataFrame"); assumed spatial domain that can be set by the user or it will be estimated by MaxEnt (see examples below)

predicted:

object of class "RasterLayer"; contains results of prediction produced using the MaxEnt software

Methods

plotKML

signature(obj = "SpatialMaxEntOutput"): plots all MaxEnt output objects in Google Earth

Note

MaxEnt requires the maxent.jar file to be in the 'java' folder of the dismo package (see: system.file("java", package="dismo")). For more info refer to the dismo package documentation.

Author(s)

Tomislav Hengl

References

Hijmans, R.J, Elith, J., (2012) Species distribution modeling with R. CRAN, Vignette for the dismo package, 72 p.

dismo package (http://CRAN.R-project.org/package=dismo)

rgbif package (http://CRAN.R-project.org/package=rgbif)

See Also

plotKML-method, dismo::maxent, rgbif::taxonsearch, GSIF::MaxEnt

[Package plotKML version 0.5-6 Index]