SpatialMetadata-class {plotKML}	R Documentation

A class for spatial metadata

Description

A class containing spatial metadata in the Federal Geographic Data Committee (FGDC) Content Standard for Digital Geospatial Metadata.

Slots

xml:

object of class "XMLInternalDocument"; a metadata slot

field.names:

object of class "character"; corresponding metadata column names

palette:

object of class "sp.palette"; contains legend names and colors

sp:

object of class "Spatial"; bounding box and projection system of the input object

Methods

summary

signature(obj = "SpatialMetadata"): summarize object

GetPalette

signature(obj = "SpatialMetadata"): get only the color slot

GetNames

signature(obj = "SpatialMetadata"): get metadata field names

Author(s)

Tomislav Hengl

See Also

spMetadata, metadata2SLD-methods

Examples


## Not run: 
data(eberg)
library(sp)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
names(eberg)
# add metadata:
eberg.md <- spMetadata(eberg, xml.file=system.file("eberg.xml", package="plotKML"), 
    Target_variable="SNDMHT_A")
p <- GetPalette(eberg.md)
str(p)
x <- summary(eberg.md)
str(x)

## End(Not run)

[Package plotKML version 0.3-3 Index]