A class for spatial predictions produced using gstat package

A class containing input and output maps generated through the process of geostatistical mapping. Object of this class can be directly visualized in Google Earth by using the plotKML-method.

Slots

variable:

object of class "character"; variable name

observed:

object of class "SpatialPointsDataFrame" (must be 2D); see sp::SpatialPointsDataFrame

regModel.summary:

contains the summary of the regression model

vgmModel:

object of class "data.frame"; contains the variogram parameters passed from gstat

predicted:

object of class "SpatialPixelsDataFrame"; see sp::SpatialPixelsDataFrame

validation:

object of class "SpatialPointsDataFrame" containing results of validation

Methods

plot

signature(x = "SpatialPredictions"): spatial predictions, regression model (observed vs predicted), original variogram and variogram for residuals

plotKML

signature(obj = "SpatialPredictions"): plots all objects in Google Earth

summary

signature(obj = "SpatialPredictions"): summarize object by showing the mapping accuracy (cross-validation) and the amount of variation explained by the model

References

Hengl, T. (2009) A Practical Guide to Geostatistical Mapping, 2nd Edt. University of Amsterdam, www.lulu.com, 291 p.

Hengl, T., Nikolic, M., MacMillan, R.A., (2012) Mapping efficiency and information content. International Journal of Applied Earth Observation and Geoinformation, special issue Spatial Statistics Conference. doi:10.1016/j.jag.2012.02.005

Note

"SpatialPredictions" saves results of predictions for a single target variable, which can be of type numeric or factor. Multiple variables can be combined into a list.

Author

Tomislav Hengl

See also