SpatialPredictions-class {plotKML}	R Documentation

A class for spatial predictions produced using gstat package

Description

A class containing input and output maps generated through the process of geostatistical mapping. Object of this class can be directly visualized in Google Earth by using the plotKML-methods.

Slots

variable:

object of class "character"; variable name

observed:

object of class "SpatialPointsDataFrame" (must be 2D); see sp::SpatialPointsDataFrame

glm:

object of class "glm"; contains the summary of the regression model

vgmModel:

object of class "data.frame"; contains the variogram parameters passed from gstat

predicted:

object of class "SpatialPixelsDataFrame"; see sp::SpatialPixelsDataFrame

validation:

object of class "SpatialPointsDataFrame" containing results of validation

Methods

plotKML

signature(obj = "SpatialPredictions"): plots all objects in Google Earth

summary

signature(obj = "SpatialPredictions"): summarize object by showing the mapping accuracy (cross-validation) and the amount of variation explained by the model

Note

"SpatialPredictions" saves results of predictions for a single target variable, which can be of type numeric or factor. Multiple variables can be combined into a list.

Author(s)

Tomislav Hengl

References

Hengl, T. (2009) A Practical Guide to Geostatistical Mapping, 2nd Edt. University of Amsterdam, www.lulu.com, 291 p.

Hengl, T., Nikolic, M., MacMillan, R.A., (2012) Mapping efficiency and information content. International Journal of Applied Earth Observation and Geoinformation, special issue Spatial Statistics Conference.

See Also

GSIF::fit.gstatModel, gstat::gstat-class, RasterBrickSimulations-class

Examples

 
# load observations:
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
# load grids:
data(meuse.grid)
gridded(meuse.grid) <- ~x+y
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
# fit a model:
library(GSIF)
omm <- fit.gstatModel(observations = meuse, formulaString = om~dist, 
  family = gaussian(log), covariates = meuse.grid)
show(omm@regModel)
# produce SpatialPredictions:
om.rk <- predict(omm, predictionLocations = meuse.grid)
# overview of the prediction process:
show(om.rk)
## Not run: # plot the whole geostatical mapping project in Google Earth:
plotKML(om.rk)

## End(Not run)

[Package plotKML version 0.2-3 Index]