SpatialSamplingPattern-class {plotKML}	R Documentation

A class for spatial samples produced using various spsample methods

Description

A class containing input and output objects generated by some sampling optimisation algorithm. Objects of this type can be directly visualized in Google Earth by using the plotKML-method.

Slots

method:

object of class "character"; sampling optimisation method

pattern:

object of class "SpatialPoints"; sampling points

sp.domain:

object of class "SpatialPolygonsDataFrame"; spatial domain / strata

Methods

plotKML

signature(obj = "SpatialSamplingPattern"): plots generated sampling plan in Google Earth

Author(s)

Tomislav Hengl

See Also

plotKML-method, spcosa::spsample, plotKML-method

[Package plotKML version 0.3-3 Index]