SpatialSamplingPattern-class {plotKML}	R Documentation

A class for spatial samples produced using various spsample methods

Description

A class containing input and output objects generated by some sampling optimisation algorithm. Objects of this type can be directly visualized in Google Earth by using the plotKML-method.

Slots

method:

object of class "character"; sampling optimisation method

pattern:

object of class "SpatialPoints"; sampling points

sp.domain:

object of class "SpatialPolygonsDataFrame"; spatial domain / strata

Methods

plotKML

signature(obj = "SpatialSamplingPattern"): plots generated sampling plan in Google Earth

Author(s)

Tomislav Hengl

See Also

plotKML-method, spcosa::spsample, plotKML-method

Examples

 
# read a polygon map:
library(rgdal)
shpFarmsum <- readOGR(dsn = system.file("maps", package = "spcosa"), layer = "farmsum")
# stratified random sampling
library(spcosa)
# stratify `Farmsum' into 50 strata
myStratification <- stratify(shpFarmsum, nStrata = 50)
# sample two sampling units per stratum
mySamplingPattern <- spsample(myStratification, n = 2)
# plot the resulting sampling pattern
plot(myStratification, mySamplingPattern)
# attach the correct proj4 string:
library(RCurl)
urlExists = url.exists("http://spatialreference.org/ref/sr-org/6781/proj4/")
if(urlExists){
  nl.rd <- getURL("http://spatialreference.org/ref/sr-org/6781/proj4/")
  proj4string(mySamplingPattern@sample) <- CRS(nl.rd) 
  # prepare spatial domain (polygons):
  #sp.domain <- as.SpatialPolygons.SpatialPixels(myStratification@cells)
  sp.domain <- as(myStratification@cells, "SpatialPolygons")
  sp.domain <- SpatialPolygonsDataFrame(sp.domain, 
    data.frame(ID=as.factor(myStratification@stratumId)), match.ID = FALSE)
  proj4string(sp.domain) <- CRS(nl.rd) 
  # create new object:
  mySamplingPattern.ssp <- new("SpatialSamplingPattern", method = class(mySamplingPattern), 
    pattern = mySamplingPattern@sample, sp.domain = sp.domain)
}

[Package plotKML version 0.3-0 Index]