aesthetics {plotKML}	R Documentation

Plotting aesthetics parameters

Description

Parses various object parameters / columns to KML aesthetics: size of the icons, fill color, labels, elevations ...

Usage


kml_aes(obj, ...)

Arguments

obj
space-time object for plotting

...
other arguments

Details

Valid aesthetics: colour = "black", fill = "white", shape, whitening, alpha, width = 1, labels, altitude = 0, size, balloon = FALSE. Specific features (target variables and the connected hot-spots) can be emphasized by using two or three graphical parameters for the same variable. See plotKML package homepage / vignette for more examples.

Author(s)

Pierre Roudier

See Also

kml-methods

[Package plotKML version 0.2-3 Index]