bigfoot {plotKML}	R Documentation

Bigfoot reports (USA)

Description

2984 observations of bigfoot (with attached dates). The data has been obtained from the BigFoot Research Organization (BFRO) website. The BFRO reports generally consist of a description of the event and where it occurred, plus the quality classification. For more info about species distribution modelling methods using the BFRO reports see Lozier et al. (2009).
The environmental predictors in the USAWgrids data set represent typical gridded maps used in species distribution modelling projects.

Usage

data(bigfoot)

Format

The bigfoot data frame contains the following columns:

Lon

a numeric vector; x-coordiante / longitude in the WGS84 system

Lat

a numeric vector; y-coordinate / latitude in the WGS84 system

NAME

name assigned by the observer (usually referent month / year)

DATE

'POSIXct' class vector

TYPE

confidence levels; according to the BFRO website: "Class A" reports involve clear sightings in circumstances where misinterpretation or misidentification of other animals can be ruled out with greater confidence; "Class B" and "Class C" reports are less credible.

The USAWgrids data frame (46,018 pixels; Washington, Oregon, Nevada and California state) contains the following columns:

globedem

a numeric vector; elevations from the ETOPO1 Global Relief Model

nlights03

an integer vector; lights at night image for 2003 (Version 2 DMSP-OLS Nighttime Lights Time Series)

sroads

a numeric vector; distance to main roads and railroads (National Atlas of the United States)

gcarb

a numeric vector; Global Biomass Carbon Map (New IPCC Tier-1 Global Biomass Carbon Map for the Year 2000)

dTRI

a numeric vector; density of pollutant releases (North American Pollutant Releases and Transfers database)

twi

a numeric vector; Topographic Wetness Index based on the globedem

states

an integer vector; USA states

globcov

land cover classes based on the MERIS FR images (GlobCover Land Cover version V2.2)

s1

a numeric vector; x-coordinates in the Albers equal-area projection system

s2

a numeric vector; y-coordinates in the Albers equal-area projection system

Note

According to the Time.com, a team of a dozen-plus experts from as far afield as Canada and Sweden have proclaimed themselves 95 percent certain of the mythical animal's existence on Kemerovo region territory some 3,000 kilometers east of Moscow (announced at the Tashtagol conference in 2011). Read more: http://newsfeed.time.com/2011/10/11/scientists-95-sure-bigfoot-lives-in-russian-tundra

Author(s)

Tomislav Hengl

References

Lozier, J.D., Aniello, P., Hickerson, M.J., (2009) Predicting the distribution of Sasquatch in western North America: anything goes with ecological niche modelling. Journal of Biogeography, 36(9):1623-1627.

BigFoot Research Organization (http://www.bfro.net)

Examples


## Load the BFRO records:
data(bigfoot)
aea.prj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
coordinates(bigfoot) <- ~Lon+Lat
proj4string(bigfoot) <- CRS("+proj=latlon +datum=WGS84")
bigfoot.aea <- spTransform(bigfoot, CRS(aea.prj))
## Covariates:
data(USAWgrids)
coordinates(USAWgrids) <- ~s1+s2
gridded(USAWgrids) <- TRUE
proj4string(USAWgrids) <- CRS(aea.prj)
## Not run: 
data(SAGA_pal)
pnts <- list("sp.points", bigfoot.aea, pch="+", col="yellow")
spplot(USAWgrids[2], col.regions=rev(SAGA_pal[[3]]), sp.layout=pnts)
## Run species distribution modelling:

## End(Not run)

[Package plotKML version 0.1-4 Index]