check_projection {plotKML}	R Documentation

Extracts the proj4 parameters and checks if the projection matches the referent CRS

Description

Function parse_proj4 gets the proj4 string from a space-time object and check_projection checks if the input projection is compatible with the referent projection system. The referent system is by default the 'latlon' projection with 'WGS84' datum (KML-compatible coordinates).

Usage

check_projection(obj, logical = TRUE, ref_CRS = get("ref_CRS", envir = plotKML.opts))

Arguments

obj 	 A Spatial* or Raster* object.

logical 	 Logical; if TRUE, a logical value is returned. If FALSE, an error is thrown if the test failed.

ref_CRS 	 The referent coordinate system.

Details

A cartographic projection is KML compatible if: (a) geographical coordinates are used, and (b) that relate to the WGS84 ellispoid ("+proj=longlat +datum=WGS84"). You can also set your own local referent projection system by specifiying plotKML.env(ref_CRS=...).

Warning

obj needs to have a proper proj4 string (CRS), otherwise check_projection will not run.

Author(s)

Pierre Roudier (pierre.roudier@gmail.com), Tomislav Hengl (tom.hengl@wur.nl), and Dylan Beaudette (debeaudette@ucdavis.edu)

References

WGS84 (http://spatialreference.org/ref/epsg/4326/)

See Also

reproject, rgdal::CRS-class

Examples


data(eberg)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
check_projection(eberg)
# not yet ready for export to KML;
parse_proj4(proj4string(eberg))
eberg.geo <- reproject(eberg)
check_projection(eberg.geo)
# now ready for export;

[Package plotKML version 0.1-2 Index]