count.GridTopology {plotKML}	R Documentation

Counts the number of occurrences of a list of vector object over a GridTopology

Description

Counts the number of occurrences of a vector object over a "GridTopology" for a list of vector objects (usually multiple realizations of the same process).

Usage


count.GridTopology(x, vectL, ...)

Arguments

x
object of type "GridTopology"

vectL
list of vectors of class "SpatialPoint*", "SpatialLines*" or "SpatialPolygons*" (equiprobable realizations of the same process)

...
(optional) arguments passed to the lower level functions

Author(s)

Tomislav Hengl

See Also

SpatialVectorsSimulations-class, vect2rast

[Package plotKML version 0.4-7 Index]