dailymeteo {plotKML}	R Documentation

Daily temperatures and precipitation for Croatia for year 2008

Description

The daily measurements of temperature (thermometers) and precipitation (rain gauges) for year 2008 obtained from the Croatian National Meteorological Service. All data is projected in the UTM zone 33N system.
HRtemp08 contains 56,608 measurements of temperature (159 stations by 365 days); HRprec08 contains 175,059 measurements of precipitation sums (489 stations by 365 days); LST is a spatial sub-sample (Istra region) of 46 time series of MODIS LST images (estimated Land Surface Temperature in degrees C) at 1 km resolution prepared as a "RasterBrick" object. The "zvalue" slot in the LST object is the dateTime column expressed as yyyy-mm-ddThh:mm:sszzzzzz, where T is the separator between the date and the time, and the time zone is either Z (for UTC) or zzzzzz, which represents ±hh:mm in relation to UTC. The temporal support of these images is 8-days.

Usage

data(HRtemp08)

Format

The HRtemp08 data frame contains the following columns:

NAME

name of the meteorological station

Lon

a numeric vector; x-coordiante / longitude in the WGS84 system

Lat

a numeric vector; y-coordinate / latitude in the WGS84 system

DATE

'Date' class vector

TEMP

daily temperature measurements in degree C

Note

The precision of the temperature readings in HRtemp08 is tenth of degree C. On most climatological stations temperature is measured three times a day, at 7 a.m., 1 p.m. and 9 p.m. The daily mean can be calculated as a weighted average. The precipitation estimates in mm (HRprec08) are collected in a bottle within the rain gauge and readings are usually manual by an observer at 7 a.m. The precipitation collected in the morning refer to the precipitation for previous 24 hours.
The time series of 46 day-time and night-time 8-day composite LST images (MOD11A2 product bands 1 and 5) was obtained from the NASA's FTP server (ftp://e4ftl01u.ecs.nasa.gov). The 8-day composite images were created by patching together images from a period of ±4 days, so that the proportion of clouds can be reduced to a minimum.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl) and Melita Percec Tadic (melita@cirus.dhz.hr)

References

Hengl, T., Heuvelink, G.B.M., Percec Tadic, M., Pebesma, E., (2011) Spatio-temporal prediction of daily temperatures using time-series of MODIS LST images. Theoretical and Applied Climatology.

http://spatial-analyst.net/book/HRclim2008

Examples


data(HRtemp08)
data(HRprec08)
data(LST)
## Not run: 
## Plot temperatures, precipitation and MODIS images in Google Earth:

## End(Not run)

[Package plotKML version 0.1-2 Index]