geopath {plotKML}	R Documentation

Geopath — shortest trajectory line between two geographic locations

Description

Derives a SpatialLines class object showing the shortest path between geographic locations and based on the Haversine Formula for Great Circle distance.

Usage


geopath(lon1, lon2, lat1, lat2, ID, n.points, print.geo = FALSE)

Arguments

lon1 	 longitude coordinate of the first point

lon2 	 longitude coordinate of the second point

lat1 	 latitude coordinate of the first point

lat2 	 latitude coordinate of the second point

ID 	 point ID vector

n.points 	 number of intermediate points

print.geo 	 prints the distance and bearing

Details

The formula to derive number of points between the start and end point is: round(sqrt(distc)/sqrt(2), 0) where distc is the Great Circle Distance.

Value

Bearing is expressed in degrees from north. Distance is expressed in kilometers (Great Circle Distance).

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

References

fossil package (http://CRAN.R-project.org/package=fossil)

Haversine formula from Math Forums (http://mathforum.org/dr.math/)

See Also

kml_layer.SpatialLines, kml_layer.STIDFtraj, fossil::earth.bear

Examples


ams.ny <- geopath(lon1=4.892222, lon2=-74.005973, lat1=52.373056, lat2=40.714353, print.geo=TRUE)
kml(ams.ny)

[Package plotKML version 0.1-2 Index]