grid2poly {plotKML}	R Documentation

Converts Gridded map to a polygon map

Description

This function converts a SpatialGridDataFrame object to a polygon map with each available grid node represented with a polygon. To allow further export to KML, it will by default convert any projected coordinates to the lat-lon system (geographic coordinates in the WGS84 system).

Usage


grid2poly(obj, var.name, reproject = TRUE, tmp.file = TRUE,  
method = c("raster", "sp", "RSAGA")[1], saga_lib = "shapes_grid", saga_module = 3, ... )

Arguments

obj 	 SpatialGridDataFrame object

var.name 	 target variable column name

reproject 	 logical; reproject coordinates to lat lon system?

method 	 decide to convert grids to polygons either using "sp", "raster" or "RSAGA" packages

saga_lib 	 string; SAGA GIS library name

saga_module 	 SAGA GIS module number; see ?rsaga_get_modules for more details

... 	 additional arguments that can be parsed to the rasterToPolygons command

Details

This operation is not recommended for large grids (>>10e4 pixels). Consider splitting large input grids into tiles before running grid2poly. Conversion of large grids to polygons is more efficient by using SAGA GIS (method = "RSAGA").

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

See Also

vect2rast, raster::rasterToPolygons

Examples


data(eberg_grid)
data(SAGA_pal)
## Not run: 
system.time(x <- grid2poly(eberg_grid, "DEM", method = "raster"))
system.time(x <- grid2poly(eberg_grid, "DEM", method = "sp"))
system.time(x <- grid2poly(eberg_grid, "DEM", method = "SAGA"))
# spplot (plotting large polygons in R not a good idea):
spplot(x, col.regions=SAGA_pal[[1]])
# plotKML:
kml(x, colour = DEM)
## End(Not run)

[Package plotKML version 0.1-2 Index]