kml {plotKML}	R Documentation

Write any Spatial*-derived or Raster* object to a KML file

Description

Writes any Spatial* object (from the sp package) or Raster* object (from the raster package) to a KML file. Various aesthetics, i.e. ways to represent target variables, can be set via colour, transparency, size, width, shape arguments. Their availability depends on the class of the object to plot.

Usage

kml(obj, folder.name, file.name, overwrite = TRUE, kmz = FALSE, ...)

Arguments

obj 	 object inheriting from the Spatial* or the Raster* classes

folder.name 	 folder name text

file.name 	 output file name (*.kml)

overwrite 	 logical; overwrites the existing file

kmz 	 logical; specifies whether to compress the kml file

... 	 additional aesthetics arguments (see details below)

Value

A KML file. By default the object name with ".kml" extension.

Author(s)

Pierre Roudier, Tomislav Hengl and Dylan Beaudette

See Also

kml_open, kml_close, kml_compress

Examples


# Plotting a SpatialPointsDataFrame object
data(eberg)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
## Not run: 
# Simple plot
kml(eberg, file = "eberg-0.kml")
# Plot using aesthetics
kml(eberg, colour = SAND, size = CLAY, alpha = 0.75, file = "eberg-1.kml")
## End(Not run)
# Plotting a SpatialPixelsDataFrame or a Raster* object
data(eberg_grid)
## Not run: 
kml(eberg_grid, colour = DEM, file = "eberg_grid-0.kml")

## End(Not run)

[Package plotKML version 0.1-2 Index]