kml-methods {plotKML}	R Documentation

Write to a KML file

Description

Writes any Spatial* object (from the sp package) or Raster* object (from the raster package) to a KML file via the plotKML.fileIO environment. Various aesthetics parameters can be set via colour, alpha, size, shape arguments. Their availability depends on the class of the object to plot.

Usage

kml(obj, ...)

Arguments

obj
object inheriting from the Spatial* or the Raster* classes

...
additional aesthetics arguments (see details below)

Details

To kml you can also pass folder.name, file.name (output file name *.kml), overwrite (logical; overwrites the existing file) and kmz (logical; specifies whether to compress the kml file) arguments.

Value

A KML file. By default parses the object name and adds a ".kml" extension.

Author(s)

Pierre Roudier, Tomislav Hengl and Dylan Beaudette

See Also

kml_open, kml_aes, kml_close, kml_compress

Examples


# Plotting a SpatialPointsDataFrame object
data(eberg)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
## Not run: 
# Simple plot
kml(eberg, file = "eberg-0.kml")
# Plot using aesthetics
kml(eberg, colour = SAND, size = CLAY, alpha = 0.75, file = "eberg-1.kml")

## End(Not run)
# Plotting a SpatialPixelsDataFrame or a Raster* object
data(eberg_grid)
coordinates(eberg_grid) <- ~x+y
gridded(eberg_grid) <- TRUE
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
## Not run: 
kml(eberg_grid, colour = DEM, file = "eberg_grid-0.kml")

## End(Not run)

[Package plotKML version 0.1-6 Index]