kml_compress {plotKML}	R Documentation

Compress a KML file with auxiliary files

Description

Compresses the KML file toghether with the auxiliary files (images, models, textures) using the default ZIP program.

Usage


kml_compress(file.name, zip = "", files = "", rm = FALSE, ...)

Arguments

file.name
KML file name

zip
(optional) location of an external ZIP program

files
a character vector specifying the list of auxiliary files

rm
logical; specify whether to remove temporary files

...
other kml arguments

Details

The KMZ file can carry the model files (.dae), textures and ground overlay images. For practical purposes, we recommend that you, instead of compressing the images together with the KML file, consider serving the ground overlay images via a server i.e. as network links.
If no internal ZIP program exists, the function looks for the system ZIP program:

Sys.getenv("R_ZIPCMD", "zip")

External ZIP program can also be specified via the zip argument.

Author(s)

Pierre Roudier, Tomislav Hengl and Dylan Beaudette

References

KMZ description (http://code.google.com/apis/kml/documentation/)

See Also

kml-methods, kml_open

Examples


## Not run: # get image info from Wikimedia:
imagename = "Soil_monolith.jpg"
x1 <- getWikiMedia.ImageInfo(imagename)
sm <- spPhoto(filename = x1$url$url, exif.info = x1$metadata)
kml_open("sm.kml")
kml_layer(sm, method="monolith")
kml_close("sm.kml")
# compress together both the KML file and the auxiliary COLLADA file:
kml_compress("sm.kml", files="Soil_monolith_jpg.dae")

## End(Not run)

[Package plotKML version 0.3-0 Index]