kml_metadata {plotKML}	R Documentation

Add metadata table to the active layer

Description

Adds a selection of metadata to the description box of an active layer.

Usage


kml_metadata(obj, sel, fix.enc = TRUE, cwidth = 150, twidth = 500,
    full.names = "", delim.sign = "_", asText = FALSE)

Arguments

obj 	 object of class "SpatialMetadata"

sel 	 a selection of metadata fields (see below)

fix.enc 	 logical; species whether the text in the XML file should be converted to UTF8

cwidth 	 html column width for the field names

twidth 	 html total table width

full.names 	 (optional) character vector; a list of full metadata names (field names)

delim.sign 	 delimiter symbol

asText 	 logical; return the output as XML or characters

Details

The kml_metadata function, by default, prints out only a number of selected metadata fields: "Citation_title", "Abstract", "Object_Count", "Beginning_Date", "Ending_Date", "Data_Order_URL", "Other_Citation_Details", "Citation_URL", "Data_Set_Credit", "Data_Distributing_Organization", "Format_Information_Content", "Native_Data_Set_Environment". See data(mdnames) for a complete list of metadata fields.

Author(s)

Tomislav Hengl

See Also

spMetadata

[Package plotKML version 0.1-2 Index]