kml_open {plotKML}	R Documentation

Opens / closes a KML file connection

Description

Opens a KML file in write mode and initiates the KML header. The same file connection is further accessible by other kml_*() functions such as kml_layer() and kml_close(). This allows creation of customized multi-layered KML files.

Usage


kml_open(file.name, folder.name = file.name,
  open = TRUE, overwrite = TRUE, kml_xsd, ...)

Arguments

file.name 	 KML file name

folder.name 	 character string; KML folder name

overwrite 	 logical; If TRUE, "name" will be overwritten if it exists

kml_xsd 	 URL of the KML scheme to be used

... 	 other arguments

Details

See plotKML package homepage / vignette for more examples.

Author(s)

Pierre Roudier and Dylan Beaudette

See Also

kml, kml.Spatial

Examples


data(eberg_grid)
kml_open("eberg_grid_DEM.kml")
kml_layer(eberg_grid, colour=DEM)
kml_close("eberg_grid_DEM.kml")
# result:
xmlRoot(xmlTreeParse("eberg_grid_DEM.kml"))[["Document"]]

[Package plotKML version 0.1-2 Index]