kml_screen {plotKML}	R Documentation

Put an image file as screen overlay

Description

Puts a selected image file (map legend or logo) on screen.

Usage


kml_screen(image.file, sname = "", position = c("UL","ML","LL","BC","LR","MR","UR","TC")[1],
  overlayXY, screenXY, xyunits = c("fraction", "pixels", "insetPixels")[1], rotation = 0, size = c(0,0) )

Arguments

image.file 	 image file to be used for screen overlay

sname 	 screen overlay name

position 	 one of the nine standard positions

overlayXY 	 manually specified tie point on the overlay image e.g. overlayXY = 'x="0" y="1"'

screenXY 	 manually specified matching tie point on the scren e.g. screenXY = 'x="0" y="1"'

xyunits 	 values of the XY units (("pixels", "fraction", or "insetPixels")

rotation 	 (optional) rotation in degrees clock-wise

size 	 size correction in x and y direction

Details

If nothing else is specified the function looks for some of the nine typical positions: "UL" (upper left), "ML" (middle left), "LL" (lower left), "BC" (bottom centre), "LR" (lower right), "MR" (middle right), "UR" (upper right), and "TC" (top centre). The x and y values can be specified in three different ways: as pixels ("pixels"), as fractions of the image ("fraction"), or as inset pixels ("insetPixels") — an offset in pixels from the upper right corner of the image. The function by default works with fractions. If you change the xyunits type, all other elements need to be expressed in the same units.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

References

KML Reference (http://code.google.com/apis/kml/documentation/)

See Also

kml

Examples


data(eberg_zones)
## Not run: 
kml_open("eberg_zones.kml")
kml_layer(eberg_zones)
logo = "http://meta.isric.org/images/ISRIC_right.png"
kml_screen(image.file = logo, position = "TC", sname = "ISRIC logo")
kml_close("eberg_zones.kml")
kml_compress("eberg_zones.kml")

## End(Not run)

[Package plotKML version 0.1-2 Index]