kml_layer.Raster {plotKML}	R Documentation

Export raster objects to KML

Description

Writes rasters to PNG images and makes a KML code (ground overlays). Works with "RasterLayer", "SpatialGrid", "SpatialPixels" and "RasterStack" class objects. Target attributes can be specified using aesthetics arguments (e.g. "colour").

Usage


kml_layer.Raster(obj, metadata = FALSE, plot.legend = TRUE, ...
  colour_scale = get("colour_scale_numeric", envir = plotKML.opts), ...)

Arguments

obj 	 object of class "RasterBrick" (a time series of images)

metadata 	 (optional) specify the metadata object

plot.legend 	 logical; specify whether a map legend should be generated automatically

... 	 additional aesthetics arguments

colour_scale 	 color palette; by default using the color scale for numeric variables

alpha 	 variable to be used to set transparency

altitude 	 if not a 3D object; variable to be used to specify altitude above ground

altitudeMode 	 altitude mode type ("clampToGround", "relativeToGround" or "absolute")

Details

Google Earth does not properly handle a 24-bit PNG file which has a single transparent color (read more at Google Earth Help). To force transparency, plotKML will try to set it using the -matte -transparent "\#FFFFFF" option in the ImageMagick convert program (ImageMagick needs to be installed separately and located using plotKML.env()).

Author(s)

Tomislav Hengl, Pierre Roudier and Dylan Beaudette

See Also

kml, kml_open, kml_layer.RasterBrick

Examples


data(eberg_grid)
data(SAGA_pal)
## Not run: 
# KML plot with a single raster:
kml(eberg_grid, colour_scale = SAGA_pal[[1]], colour = TWI)

## End(Not run)

[Package plotKML version 0.1-2 Index]