kml_layer.RasterBrick {plotKML}	R Documentation

Export a time series of rasters to KML

Description

Writes a series of images to png and creates ground overlays.

Usage


kml_layer.RasterBrick(obj, obj.title, id.name, dtime = 0, extrude = FALSE,
   z.scale = 1, metadata = FALSE, html.table = NULL)

Arguments

obj 	 object of class "RasterBrick" (a time series of images)

obj.title 	 object title that will be inserted in the folder name

dtime 	 temporal support - point or block

metadata 	 logical; should metadata be printed?

html.table 	 optional description block (html) for each GPS point (vertices)

tz = "GMT" 	 time zone

z.lim 	 upper and lower limits (unique for all maps in the time series)

Details

This method is recommended for numeric bands representing the same feature i.e. the same variable (time series of images). To export a stack of images of different type see kml_layer.Raster.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

See Also

kml, kml_open, kml_layer.Raster

Examples


data(LST)
# time dimension defined via zvalue slot:
LST@zvalue[1:2]
## Not run: 
# subset to 5 images:
LSTs <- brick(raster(LST, 1), raster(LST, 2), raster(LST, 3))
LST.ll <- reproject(LST, program="FWTools")
LST.ll@zvalue = LST@zvalue
# KML plot with block temporal support of 8 days:
kml(LST.ll, colour_scale=SAGA_pal[[1]], dtime=8*24*60^2)
## End(Not run)

[Package plotKML version 0.1-2 Index]