kml_layer.STIDFtraj {plotKML}	R Documentation

Writes a space-time trajectory object to KML

Description

Writes a space-time trajectory object to a KML file and (optional) attaches attributes.

Usage


kml_layer.STIDFtraj(obj, id.name, dtime = "", extrude = FALSE,
   start.icon, end.icon, LabelScale, z.scale = 1, metadata = NULL, html.table = NULL)

Arguments

obj 	 space-time object of class "STIDFraj" (spatio-temporal irregular data.frames trajectory)

id.name 	 trajectory ID column name

dtime 	 temporal support - point or block

extrude 	 logical; extrude GPS vertices?

start.icon 	 start icon name (3Dballyellow.png)

end.icon 	 destination icon name (golfhole.png)

LabelScale 	 the default size of icons

z.scale 	 vertical exaggeration

metadata 	 (optional) specify the metadata object

html.table 	 optional description block (html) for each GPS point (vertices)

Details

The dateTime is defined as yyyy-mm-ddThh:mm:sszzzzzz, where T is the separator between the date and the time, and the time zone is either Z (for UTC) or zzzzzz, which represents \enc±+-hh:mm in relation to UTC. For more info on how Time Stamps work see http://kml-samples.googlecode.com. If the time is measured at block support, then <TimeStamp><begin> </begin><end> </end></TimeStamp> times will be inserted. Temporal support (dt) is estimated using the xts::periodicity function (median time between observations). The begin time is then derived as t1 = t - .5 d t.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

See Also

kml, kml_open, readGPX

Examples


data(gpxbtour)
# format the time column:
gpxbtour$ctime <- as.POSIXct(gpxbtour$time, format="%Y-%m-%dT%H:%M:%SZ")
coordinates(gpxbtour) <- ~lon+lat
proj4string(gpxbtour) <- CRS("+proj=longlat +datum=WGS84")
## Not run: 
# derive distance from origin:
require(fossil)
xy <- as.list(data.frame(t(coordinates(gpxbtour))))
gpxbtour$dist.km <- sapply(xy, function(x) deg.dist(long1=x[1], lat1=x[2], long2=xy[[1]][1], lat2=xy[[1]][2]))
plot(gpxbtour$dist.km, gpxbtour$speed, type="l", xlab="Distance from origin in km", ylab="speed in km per hour")
# convert to a STIDFtraj class:
require(spacetime)
require(adehabitatLT)
gpx.ltraj <- as.ltraj(coordinates(gpxbtour), gpxbtour$ctime, id = "th")
gpx.st <- as(gpx.ltraj, "STIDFtraj")
gpx.st$speed <- gpxbtour$speed
gpx.st@sp@proj4string <- CRS("+proj=longlat +datum=WGS84")
# open KML file:
kml_open("gpxbtour_speed.kml")
# plot trajectory with speed as colour attribute:
kml_layer(obj=gpx.st, colour=speed, z.lim=c(18,32))
# close KML:
kml_close("gpxbtour_speed.kml")
kml_compress("gpxbtour_speed.kml")
## End(Not run)

[Package plotKML version 0.1-2 Index]