kml_layer.SoilProfileCollection {plotKML}	R Documentation

Export a list of soil profiles to KML

Description

Writes object of type "SoilProfileCollection" (a number of soil profiles with site and horizon data) to KML. Several attributes such as horizontal and vertical exaggeration can be passed via arguments.

Usage


kml_layer.SoilProfileCollection(obj, var.name, site_names = obj@site[,1],
  z.scale = 1, x.min = 0, x.scale, max.depth = 300, plot.points = TRUE, 
  LabelScale, IconColor = "ff0000ff", shape = paste(get("home_url", envir = plotKML.opts), "circlesquare.png", sep=""), 
  visibility = TRUE, extrude = TRUE, tessellate = TRUE, camera.distance = .01, 
  altitudeMode = "relativeToGround", tilt = 90, heading = 0, roll = 0, 
  metadata = NULL, html.table = NULL, ...)

Arguments

obj 	 object of class "SoilProfileCollection" (package aqp)

var.name 	 target column name in the horizons slot

site_names 	 site names as listed in the site table

z.scale 	 exaggeration in vertical dimension

x.min 	 offset in longitude direction (in decimal degrees)

x.scale 	 exaggeration in horizontal direction

max.depth 	 maximum height/depht of a profile in cm

plot.points 	 logical; specifies whether to plot horizon centres with attribute values

camera.distance 	 distance from a profile in arc degrees

altitudeMode 	 by default relativeToGround

tilt 	 angle between the direction of the LookAt position and the normal to the surface of the earth

heading 	 orientation towards north

roll 	 rotation about the y axis

... 	 additional style arguments

Details

Horizon depths are typically expressed in cm, hence the default exaggeration factor (z.scale) is 10. It is highly recommended to turn off the terrain layer in Google Earth, otherwise Google Earth will deform the plots in areas of high relief.

Author(s)

Tomislav Hengl, Dylan Beaudette and Pierre Roudier

References

Algorithms for Quantiative Pedology (http://CRAN.r-project.org/package=aqp)

See Also

kml_layer.SpatialPhotoOverlay

Examples


install.packages("aqp", repos="http://R-Forge.R-project.org")
library(aqp)
data(ca630)
# Promote to SoilProfileCollection
ca <- join(ca630$lab, ca630$site, type='inner')
depths(ca) <- pedon_key ~ hzn_top + hzn_bot
# extract site data
site(ca) <- ~ mlra + ssa + lon + lat + cntrl_depth_to_top + cntrl_depth_to_bot + sampled_taxon_name
# generate SpatialPoints
coordinates(ca) <- ~ lon + lat
# assign CRS data
proj4string(ca) <- "+proj=latlong +datum=NAD83"
## Not run: 
# plot changes in base saturation by sum of cations method (pH 8.2):
kml(ca, method = "depth_function", file.name = "ca_bs_8_2.kml", var.name="bs_8.2", balloon = TRUE)
# plot changes in cation exchange capacity by sum of cations method (pH 8.2):
kml(ca, file.name = "ca_CEC8_2.kml", var.name="CEC8.2", IconColor = "#ff009000")
# plot soil profile as 'block':
kml(ca, file.name = "ca_CEC8_2_block.kml", var.name="CEC8.2", balloon = TRUE)

## End(Not run)

[Package plotKML version 0.1-2 Index]