kml_layer.SpatialLines {plotKML}	R Documentation

Exports objects of type SpatialLines* to KML

Description

Writes object of type SpatialLines* to KML with several aesthetics arguments.

Usage


kml_layer.SpatialLines(obj, extrude = FALSE, z.scale = 1, spMetadata = NULL, 
  html.table = NULL, ...)

Arguments

obj 	 object of class "RasterBrick" (a time series of images)

spMetadata 	 (optional) specify the metadata object

html.table 	 optional description block (html) for each GPS point (vertices)

... 	 additional style arguments (see aesthetics)

labels 	 variable to be used to attach labels to spatial lines

colour 	 variable to be used to color spatial lines

width 	 variable to be used to specify the width of spatial lines

altitude 	 if not a 3D object; variable to be used to specify altitude above ground

altitudeMode 	 altitude mode type (clampToGround, relativeToGround or absolute)

Details

Only colour and width (aesthetics) are recommended when visualizing SpatialLines.

Author(s)

Pierre Roudier (pierre.roudier@landcare.nz), Tomislav Hengl (tom.hengl@wur.nl), Dylan Beaudette (debeaudette@ucdavis.edu)

See Also

kml, kml_open, kml_layer.SpatialPolygons

Examples


data(eberg_contours)
data(SAGA_pal)
names(eberg_contours)
## Not run: 
# KML plot with elevations used as 'colour' argument:
kml(eberg_contours, colour_scale=SAGA_pal[[1]], colour=Z)
## End(Not run)

[Package plotKML version 0.1-2 Index]