kml_layer.SpatialPhotoOverlay {plotKML}	R Documentation

Exports objects of type SpatialPhotoOverlay to KML

Description

Writes object of type SpatialPhotoOverlay to KML together with a COLLADA 3D model file (optional).

Usage


kml_layer.SpatialPhotoOverlay(obj, method = c("PhotoOverlay", "monolith")[1], 
  PhotoOverlay.shape = obj@PhotoOverlay$shape, href = obj@filename, coords,
  dae.name = "", heading = obj@PhotoOverlay$heading, tilt = obj@PhotoOverlay$tilt,
  roll = obj@PhotoOverlay$roll, near = obj@PhotoOverlay$near, range = obj@PhotoOverlay$range,
  leftFov = obj@PhotoOverlay$leftFov, rightFov = obj@PhotoOverlay$rightFov, bottomFov = obj@PhotoOverlay$bottomFov, 
  topFov = obj@PhotoOverlay$topFov, altitudeMode = "clampToGround", block.size = 100,
  max.depth = 300, scale.x = 1, scale.y = 1, scale.z = 1, refreshMode = "once", ...)

Arguments

obj 	 object of class "SpatialPhotoOverlay" (a photograph with spatial coordinates, metadata and orientation)

method 	 visualization type: either "PhotoOverlay" or "monolith"

PhotoOverlay.shape 	 PhotoOverlay shape value (KML)

href 	 location of the image file

dae.name 	 (optional) COLLADA 3D model file name (without the extension)

... 	 a list of PhotoOverlay related arguments

block.size 	 width of the block (100 m by default)

max.depth 	 300 m by default

html.table 	 (optional) specify the description block (html) for each point

... 	 other additional arguments

Details

The default widht and height (100 m and 300 m) were selected based on empirical testing (level of detail in the background imagery in Google Earth). User specified coordinates can be passed via the cords argument. For more info see makeCOLLADA.rectangle.

Author(s)

Tomislav Hengl

References

KML Reference (http://code.google.com/apis/kml/documentation/kmlreference.html)

See Also

spPhoto, getWikiMedia.ImageInfo

Examples


## Not run: 
imagename = "Soil_monolith.jpg"
x1 <- getWikiMedia.ImageInfo(imagename)
sm <- spPhoto(filename = x1$url$url, exif.info = x1$metadata)
kml_open("sm.kml")
kml_layer(sm, method="monolith")
kml_close("sm.kml")
kml_compress("sm.kml", files="Soil_monolith_jpg.dae")

## End(Not run)

[Package plotKML version 0.1-2 Index]