kml_layer.SpatialPixels {plotKML}	R Documentation

Writes SpatialPixels or SpatialGrid objects to KML

Description

Writes sp classes "SpatialGrid" or "SpatialPixels" to PNG images and makes a KML document (ground overlays). Target attributes can be specified using aesthetics arguments (e.g. "colour").

Usage


kml_layer.SpatialPixels(obj, raster_name, plot.legend = TRUE, metadata = NULL, ...)

Arguments

obj
object of class "RasterLayer", "SpatialPixelsDataFrame" or "SpatialGridDataFrame"

plot.legend
logical; specify whether a map legend should be generated automatically

metadata
(optional) specify the metadata object

raster_name
(optional) specify the output file name (PNG)

...
additional aesthetics arguments

Details

Google Earth does not properly handle a 24-bit PNG file which has a single transparent color (read more at Google Earth Help). To force transparency, plotKML will try to set it using the -matte -transparent "\#FFFFFF" option in the ImageMagick convert program (ImageMagick needs to be installed separately and located using plotKML.env()). The PNG export uses the 'cairographics', which will never use a palette and normally creates a larger 32-bit ARGB file, but then always allows transparancy.

Author(s)

Tomislav Hengl, Pierre Roudier and Dylan Beaudette

See Also

kml-methods, kml_open, kml_layer.Raster

Examples


data(eberg_grid)
coordinates(eberg_grid) <- ~x+y
gridded(eberg_grid) <- TRUE
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
data(SAGA_pal)
## Not run: 
# KML plot with a single raster:
kml(eberg_grid, colour_scale = SAGA_pal[[1]], colour = TWISRT6) 
# factor-type variable:
kml(eberg_grid, colour_scale = SAGA_pal[[1]], colour = LNCCOR6)

## End(Not run)

[Package plotKML version 0.3-0 Index]