metadata2SLD {plotKML}	R Documentation

Create a Styled Layer Description (SLD) file

Description

Creates a Styled Layer Description (SLD) file, that can be attached to a spatial layer contributed to GeoServer. This basically writes the "sp.pallete" object (legend entries, titles and colors) to an external file.

Usage


metadata2SLD(obj, obj.name = deparse(substitute(obj)), 
  sld.file = set.file.extension(obj.name, ".sld"), 
  Citation_title = xmlValue(obj@xml[["//title"]]), 
  ColorMap_type = "intervals", opacity = 1)

Arguments

obj 	 object of class "SpatialMetadata"

obj.name 	 object name to be inserted in the description tab

sld.file 	 external SLD file

Citation_title 	 full description of the layer

ColorMap_type 	 html total table width

opacity 	 optional opacity index (0-1)

Details

The structure of the SLD file is determined by the object class (Point, Polygon, SpatialPixels).

Author(s)

Tomislav Hengl

See Also

spMetadata

[Package plotKML version 0.1-2 Index]