normalizeFilename {plotKML}	R Documentation

Normalize filename string

Description

Remove all reserved characters from the file name.

Usage


normalizeFilename(x, format = c("default", "8.3")[1], fix.encoding = TRUE, sub.sign = "_")

Arguments

x 	 input character

format 	 target format (standard or the short 8.3 file name)

fix.encoding 	 logical; specifies whether to fix the encoding

sub.sign 	 substitution symbol

Details

This function removes all reserved characters: (less than), (greater than), (colon), (double quote), (forward slash), (backslash), (vertical bar or pipe), (question mark), (asterisk), and empty spaces, from the file name. This is important when writing a list of objects to KML as it prevents from creating erroneous file names.

Author(s)

Tomislav Hengl

See Also

utils::shortPathName

Examples


normalizeFilename("name[%].txt")
normalizeFilename("name .txt")

[Package plotKML version 0.1-2 Index]