plotKML-methods {plotKML}	R Documentation

Methods for plotting results of spatial analysis in Google Earth

Description

The method writes inputs and outputs of spatial analysis (a list of point, gridded and polygon data usually) to KML and opens the KML file in Google Earth (or any other default package used to view KML/KMZ files).

Usage

plotKML(obj, ... )

Arguments

obj
input object of specific class; usually a composite object containing both inputs and outputs of analysis

...
(optional) arguments passed to the lower level functions

Details

This is a generic function to plot R objects that contain both inputs and outputs of spatial analysis. The resulting plots (referred to as ‘views’) are usually cartographically complete as they contain legends, and data and model descriptions. In principle, plotKML works only with complex objects such as "SpatialPredictions", "SpatialLinesSimulations", "RasterBrickSimulations", "RasterBrickTimeSeries", "SpeciesDistributionMap" and similar. To further customize visualizations consider combining the lower level functions kml_open, kml_close, kml_compress, kml_screen into your own wrapper function.

Note

Writting large spatial objects via plotKML can be time consuming. Please refer to the package documentation for more information.

See Also

plotKML-package

[Package plotKML version 0.1-5 Index]