readGPX {plotKML}	R Documentation

Imports GPX (GPS track) files to R

Description

Read various elements from a
*.gpx
file - metadata, waypoints, tracks and routes - and converts them to tables.

Usage

readGPX(gpx.file, metadata = TRUE, bounds = TRUE,
    waypoints = TRUE, tracks = TRUE, routes = TRUE)

Arguments

gpx.file 	 location of the gpx.file

metada 	 logical; should the metadata be imported?

bounds 	 logical; should bounding box coordinates be imported?

waypoint 	 logical; should all waypoints be imported?

track 	 logical; should all tracks be imported?

route 	 logical; should all routes be imported?

Details

Waypoint is a point of interest, or named feature on a map. Track is an ordered list of points describing a path. Route is an ordered list of waypoints representing a series of turn points leading to a destination.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

References

GPX data format (http://www.topografix.com/gpx.asp)

XML tutorial (http://www.omegahat.org/RSXML/Tour.pdf)

See Also

rgdal::readOGR, kml_layer.STIDFtraj

Examples


## Not run: 
gpxbtour <- readGPX(system.file("btour.gpx", package="plotKML"))
str(gpxbtour)
fells_loop <- readGPX("http://www.topografix.com/fells_loop.gpx")
str(fells_loop)
## End(Not run)

[Package plotKML version 0.1-2 Index]