readKML.GBIFdensity {plotKML}	R Documentation

Imports GBIF cell density records

Description

Read GBIF cell (1–degree) density record counts and converts them to a "raster" object.

Usage

readKML.GBIFdensity(kml.file, gbif.url = FALSE, silent = FALSE)

Arguments

kml.file 	 GBIF cell density file (local file or URL)

gbif.url 	 logical; species whether the cellid and taxon content information should be also imported (usually not used)

silent 	 logical; species whether the progress bar should be printed

Details

This document contains data shared through the GBIF Network — see http://data.gbif.org for more information. GBIF records are constantly updated and every map derived refers to a certain date indicated in the @zname Last update slot.
All usage of these data must be in accordance with the GBIF Data Use Agreement: http://www.gbif.org/DataProviders/Agreements/DUA.

Author(s)

Tomislav Hengl

References

GBIF cell density description (http://data.gbif.org/ws/rest/density)

See Also

readGPX

Examples


## Not run: 
kml.file <- "taxon-celldensity-2294100.kml"
## download.file(paste("http://data.gbif.org/occurrences/taxon/celldensity/", kml.file, sep=""), destfile=paste(getwd(), kml.file, sep="")) 
## this will not run (you must first accept the data usage agreeent); 
## instead, obtain the kml file via a web browser, and save it to the working directory:
r <- readKML.GBIFdensity(kml.file)
class(r)
summary(r)
image(r)
load(url("http://spatial-analyst.net/DATA/worldborders.RData"))
lines(worldborders)
# To import a list of files use e.g.:
kml.list <- list(kml.file)
r.lst <- lapply(kml.list, readKML.GBIFdensity, silent = TRUE)
# mask out missing layers (empty KML files):
mask <- !sapply(r.lst, is.null)
r.lst <- brick(r.lst[mask])

## End(Not run)

[Package plotKML version 0.1-2 Index]