reproject {plotKML}	R Documentation

Reprojects data to a given coordinate reference system

Description

This wrapper function reprojects any vector or raster spatial data to some referent coordinate system (by default: geographic coordinates on the World Geodetic System of 1984 / WGS84 datum).

Usage

reproject(obj, CRS = get("ref_CRS", envir = plotKML.opts), ..., program = "raster")

Arguments

obj 	 Spatial* or Raster* object

CRS 	 proj4string; see also rgdal::CRS-class

... 	 arguments evaluated in the context of function projectRaster from the raster package

program 	 default program for reprojecting vectors / grids; "raster" or "FWTools"

Details

In the case of raster and/or gridded maps, by selecting program = "FWTools" gdalwarp functionality is initiated. This requires that FWTools are installed and listed via paths(). FWTools includes OpenEV, GDAL, MapServer, PROJ.4 and OGDI, and can be installed on Windows and Unix-type operating systems.

Warning

obj needs to have a proper proj4 string (CRS), otherwise reproject will not run.

Author(s)

Pierre Roudier, Tomislav Hengl and Dylan Beaudette

References

Raster package (http://CRAN.R-project.org/package=raster)

FWTools (http://fwtools.maptools.org)

See Also

paths, raster::projectRaster, rgdal::spTransform

Examples


# Example with vector data
data(eberg)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
eberg.geo <- reproject(eberg)
# Example with raster data
data(eberg_grid)
## Not run: 
eberg_grid.geo <- reproject(eberg_grid)
## End(Not run)
# Using FWTools:
## Not run: 
eberg_grid.geo <- reproject(eberg_grid, program = "FWTools")
## End(Not run)
# Optional: compare processing times:
system.time(eberg_grid_ll <- reproject(eberg_grid[1]))
system.time(eberg_grid_ll <- reproject(eberg_grid[1], program="FWTools"))

[Package plotKML version 0.1-2 Index]