sp.palette-class {plotKML}	R Documentation

A class for color palette

Description

A class for color palette that can be further used to create an object of class "SpatialMetadata".

Slots

bounds:

object of class "numeric" or "character"; class boundaries

color:

object of class "character"; contains HEX colors

icons:

object of class "character"; (optional) contains symbols or URI to icons

names:

object of class "character"; class names (optional)

type:

object of class "character"; variable type

Note

Size of class boundaries (upper and lower) is 1 element larger then the size of colors and element names.

Author(s)

Tomislav Hengl

See Also

spMetadata, metadata2SLD-methods

[Package plotKML version 0.4-7 Index]