spMetadata {plotKML}	R Documentation

Generate spatial metadata

Description

The read.metadata function can be used to read the formatted metadata (.xml), prepared following to the Federal Geographic Data Committee (FGDC) Content Standard for Digital Geospatial Metadata, and convert them to a more readble format (data frame).
The spMetadata function will try to generate missing metadata (bounding box, location info, session info, metadata creator info etc) for any Spatial* object (from the sp package) or Raster* object (from the raster package). The resulting object of class "SpatialMetadata" can be used e.g. to generate a Layer description documents, that can be parsed to KML <description> tag (see kml_metadata function).

Usage

spMetadata(obj, xml.file = set.file.extension(deparse(substitute(obj)), ".xml"),  
    Citation_title, Target_variable, Attribute_Measurement_Resolution = 1,
    Attribute_Units_of_Measure = "NA", Enduser_license_URL = get("license_url", envir = plotKML.opts)
    generate.missing = TRUE, GoogleGeocode = FALSE, signif.digit = 3, colour_scale, 
    bounds, legend_names, icons, validate.schema = FALSE, ...)

Arguments

obj 	 some "Spatial" class object with "data" slot

xml.file 	 (optional) metadata file in the FGDC format

Citation_title 	 one-sentence title of the object

Target_variable 	 target variable i.e. column name in the data slot

Attribute_Measurement_Resolution 	 numeric resolution (typically estimated as half the measurement precision).

Attribute_Units_of_Measure 	 physical units of the target variable

Enduser_license_URL 	 end-user license specification (URL)

generate.missing 	 logical; species whether the program should try to generate missing metadata automatically

GoogleGeocode 	 logical; specifies whether R should try to estimate Indirect Spatial Reference using Google maps API

signif.digit 	 integer; a number of significant digits to round the numbers in the palette slot

colour_scale 	 collor scheme used to visualize this data

bounds 	 numeric vector; upper and lower limit bounds for the target variable

legend_names 	 (optional) manually inserted legend names

icons 	 (optional) manually inserted icon names for legend names

validate.schema 	 logical; specieis whether R should try to validate your xml file

... 	 additional kml() arguments (see details below)

Details

spMetadata tries to locate a metadata file in the working directory (it looks for a metadata file with the same name as the object name). If no .xml file exists, it will load the template xml file available in the system folder (system.file("FGDC.xml", package="plotKML")). The FGDC.xml file contains typical metadata entries with description and examples. For practical purposes, one metadata object in plotKML can be associated with only one variable i.e. one column in the "data" slot (specified via Target_variable). To prepare a metadata xml file for the layer of interest, consider using the FGDC.xml template file; to enter the missing fields use the Tkme program: Another editor for formal metadata, by Peter N. Schweitzer (U.S. Geological Survey).
The spMetadata function tries to automatically generate the most usefull information, so that a user can easily find out about the input data and procedures followed to generate the visualization (KML). Typical metadata entries include e.g.:

metadata[["idinfo"]][["native"]] — Session info e.g.: Produced using R version 2.12.2 (2011-02-25) running on Windows 7 x64.

metadata[["spdoinfo"]][["indspref"]] — Indirect spatial reference estimated using the Google Maps API Web Services.

metadata[["idinfo"]][["spdom"]][["bounding"]] — Bounding box in the WGS84 geographical coordinates estimated by reprojecting the original bounding box.
By default, plotKML uses the Creative Commons license, but this can be adjusted by setting the Enduser_license_URL argument above.

Value

xml 	 a metadata slot of class "XMLInternalDocument"

field.names 	 corresponding metadata column names

palette 	 formal class "sp.palette" — contain legend names and colors

sp 	 formal class "Spatial" — bounding box and projection system of the input object

Author(s)

Tomislav Hengl, Pierre Roudier, and Dylan Beaudette

References

The Federal Geographic Data Committee, (2006) FGDC Don't Duck Metadata — A short reference guide for writing quality metadata. Vers. 1, http://www.fgdc.gov/metadata/documents/MetadataQuickGuide.pdf

Content Standard for Digital Geospatial Metadata (http://www.fgdc.gov/metadata/csdgm/)

Template metadata file (http://plotkml.r-forge.r-project.org/FGDC.xml)

Tkme metadata editor (http://geology.usgs.gov/tools/metadata/tools/doc/tkme.html)

Creative Commons license (http://creativecommons.org/licenses/)

See Also

sp::Spatial, kml_open

Examples


## Not run: 
# Read metadata from the system file (convert to a more redable data.frame format):
x <- read.metadata(system.file("FGDC.xml", package="plotKML"))
str(x)
# Generate missing metadata
data(eberg)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
# No metadata file specified:
eberg.md <- spMetadata(eberg, Target_variable="SAND")
# With localy prepared metadata file:
eberg.md <- spMetadata(eberg, xml.file=system.file("eberg.xml", package="plotKML"), Target_variable="SAND")
# Print the results of validation (it needs to be able to connect to a URL with schema):
eberg.md <- spMetadata(eberg, Target_variable="SAND", validate.schema = TRUE)
# Write the metadata to a file:
saveXML(metadata(eberg.md), "eberg_SAND.xml")
# Export to SLD format:
metadata2SLD(eberg.md, "eberg.sld")
# plot the layer with the metadata:
kml(eberg, colour = SAND, metadata = eberg.md)
## End(Not run)

[Package plotKML version 0.1-2 Index]