vect2rast {plotKML}	R Documentation

Converts Points, Lines and Polygons to rasters

Description

This function converts any Spatial-PointsDataFrame, -LinesDataFrame, or -PolygonsDataFrame object to a raster map, and (optional) writes it to an external file (GDAL-supported formats). The cell size for output raster map is by default determined based on the density/size of features in the vector map.

Usage


vect2rast(obj, var.name = names(obj)[1], cell.size, bbox, file.name, ...)

Arguments

obj 	 Spatial-PointsDataFrame,-LinesDataFrame or -PolygonsDataFrame object

var.name 	 target variables in the "data" slot

cell.size 	 (optional) grid cell size in the output raster map

bbox 	 (optional) output bounding box (class "bbox") for cropping the data

file.name 	 (optional) file name to export the resulting raster map

... 	 additional arguments to be passed to the rasterize command

Details

This function basically extends the rasterize function available in the raster package. the advantage of vect2rast is that it requires no input from the user's side i.e. it automatically determines the grid cell size based on the properties of the input data set.
The optimal grid cell size is estimated based on the density/size of features in the map (nndist function in spatstat package): (a) in the case of SpatialPoints cell size is determined as half the mean distance between the nearest points; (b) in the case of SpatialLines half cell size is determined as half the mean distance between the lines; (c) in the case of polygon data cell size is determined as half the median size (area) of polygons of interest. For more details see Hengl (2006).

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

References

Hengl T., (2006) Finding the right pixel size. Computers and Geosciences, 32(9): 1283-1298.

SpatStat package (http://www.spatstat.org)

See Also

raster::rasterize, spatstat::nndist

Examples


data(eberg)
coordinates(eberg) <- ~X+Y
## Not run: 
x <- vect2rast(eberg)
image(x)
## End(Not run)

[Package plotKML version 0.1-2 Index]