vector2raster {plotKML}	R Documentation

Converts Points, Lines and Polygons to rasters

Description

This function converts any Spatial-PointsDataFrame, -LinesDataFrame, or -PolygonsDataFrame object to a raster map, and (optional) writes it to an external file (GDAL-supported formats). The cell size for output raster map is by default determined based on the density/size of features in the vector map.

Usage


vector2raster(obj, var.name = names(obj)[1], 
cell.size, bbox, file.name, ...)

Arguments

obj 	 Spatial-PointsDataFrame,-LinesDataFrame or -PolygonsDataFrame object.

var.name 	 Target variables in the "data" slot.

cell.size 	 (Optional) grid cell size in the output raster map.

bbox 	 (Optional) output bounding box (class "bbox") for cropping the data.

file.name 	 (Optional) file name to export the resulting raster map.

... 	 Additional arguments to be parsed to the rasterize command.

Details

This function basically extends the rasterize function available in the raster package. vector2raster, however, requires no input from the user's side.
The optimal grid cell size is estimated based on the density/size of features in the map: (a) in the case of SpatialPoints cell size is determined as half the mean distance between the nearest points; (b) in the case of SpatialLines half cell size is determined as half the mean distance between the lines; (c) in the case of polygon data cell size is determined as half the median size (area) of polygons of interest. For more details see Hengl (2006).

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

References

Hengl T., (2006) Finding the right pixel size. Computers and Geosciences, 32(9): 1283-1298.

See Also

raster::rasterize

Examples


data(eberg)

## Not run: 
x <- vector2raster(eberg)
image(x)

## End(Not run)

[Package plotKML version 0.1-2 Index]