Daily precipitation for Croatia for year 2008

The daily measurements of precipitation (rain gauges) for year 2008 kindly contributed by the Croatian National Meteorological Service. HRprec08 contains 175,059 measurements of precipitation sums (489 stations by 365 days).

data(HRprec08)

Format

The HRprec08 data frames contain the following columns:

NAME

name of the meteorological station

Lon

a numeric vector; x-coordiante / longitude in the WGS84 system

Lat

a numeric vector; y-coordinate / latitude in the WGS84 system

DATE

'Date' class vector

PREC

daily cummulative precipitation in mm (precipitation from the day before)

Author

Tomislav Hengl and Melita Percec Tadic

References

Testik, F.Y. and Gebremichael, M. Eds (2011) Rainfall: State of the Science. Geophysical monograph series, Vol. 191, 287 p. doi:10.1029/GM191

Zaninovic K., Gajic-Capka, M., Percec Tadic, M. et al., (2010) Klimatski atlas Hrvatske / Climate atlas of Croatia 1961-1990., 1971-2000. Zagreb, Croatian National Meteorological Service, 200 p.

AGGM book datasets (http://spatial-analyst.net/book/HRclim2008)

Note

The precipitation estimates in mm (HRprec08) are collected in a bottle within the rain gauge and readings are usually manual by an observer at 7 a.m. The precipitation collected in the morning refer to the precipitation for previous 24 hours. To project coordinates we suggest using the UTM zone 33N system as this coordinate system was used to prepare the gridded predictors.

See also

Examples

data(HRprec08)
library(sp)
if (FALSE) # subset:
prec.2008.05.01 <- HRprec08[HRprec08$DATE=="2008-05-01",]
coordinates(prec.2008.05.01) <- ~Lon+Lat
#> Error in coordinates(prec.2008.05.01) <- ~Lon + Lat: object 'prec.2008.05.01' not found
proj4string(prec.2008.05.01) <- CRS("+proj=lonlat +datum=WGS84")
#> Warning: 'lonlat' changed to 'longlat': +proj=longlat +datum=WGS84
#> Error in proj4string(prec.2008.05.01) <- CRS("+proj=lonlat +datum=WGS84"): object 'prec.2008.05.01' not found
# write to KML:
shape = "http://plotkml.r-forge.r-project.org/circle.png"
data(SAGA_pal)
kml(prec.2008.05.01, size = PREC, shape = shape, colour = PREC, 
    colour_scale = SAGA_pal[[9]], labels = PREC)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'obj' in selecting a method for function 'kml': object 'prec.2008.05.01' not found