LST {plotKML}	R Documentation

Time series of MODIS LST images (estimated Land Surface Temperature in degrees C) at 1 km resolution

Description

LST contains a spatial sub-sample (Istra region in Croatia) of 46 time series of MODIS LST images (estimated Land Surface Temperature in degrees C) at 1 km resolution. The "zvalue" slot in the "RasterBrick" object can be used as the dateTime column expressed as yyyy-mm-ddThh:mm:sszzzzzz, where T is the separator between the date and the time, and the time zone is either Z (for UTC) or zzzzzz, which represents ±hh:mm in relation to UTC. The temporal support of these images is 8-days.

Usage

data(LST)

Format

The LST data frame contains the following layers:

LST2008_01_01

8-day MODIS LST mosaick for period 2007-12-29 to 2008-01-04

LST2008_01_09

8-day MODIS LST mosaick for period 2008-01-05 to 2008-01-13

...

subsequent bands

x

a numeric vector; x-coordinate (m) in the UTM system

y

a numeric vector; y-coordinate (m) in the UTM system

Note

Time series of 46 day-time and night-time 8-day composite LST images (MOD11A2 product bands 1 and 5) was obtained from the NASA's FTP server (ftp://e4ftl01u.ecs.nasa.gov). The original 8-day composite images were created by patching together images from a period of ±4 days, so that the proportion of clouds can be reduced to a minimum.

Author(s)

Tomislav Hengl and Melita Percec Tadic

References

Hengl, T., Heuvelink, G.B.M., Percec Tadic, M., Pebesma, E., (2011) Spatio-temporal prediction of daily temperatures using time-series of MODIS LST images. Theoretical and Applied Climatology, 107(1-2): 265-277.

MODIS products (https://lpdaac.usgs.gov/lpdaac/products/modis_products_table/)

Examples


data(LST)
coordinates(LST) <- ~x+y
gridded(LST) <- TRUE
proj4string(LST) <- CRS("+proj=utm +zone=33 +datum=WGS84 +units=m")
# get the dates from the file names:
dates <- sapply(strsplit(names(LST), "LST"), function(x){x[[2]]})
LST <- brick(LST)
LST@title = "Time series of MODIS Land Surface Temperature (8-day mosaics) images"
LST@zvalue <- format(as.Date(dates, "%Y_%m_%d"), "%Y-%m-%dT%H:%M:%SZ")
data(SAGA_pal)
## Not run: 
# Plot temperatures, precipitation and MODIS images in Google Earth:
LST_ll <- reproject(LST)
# copy the dates:
LST_ll@zvalue <- LST@zvalue
# KML plot with block temporal support of 8 days:
kml(LST_ll, colour_scale=SAGA_pal[[1]], dtime=8*24*60^2)

## End(Not run)

[Package plotKML version 0.1-4 Index]