RasterBrickTimeSeries-class {plotKML}	R Documentation

A class for a time series of regular grids

Description

A class containing list of rasters, begin, end times and sample points to allow exploration of the values. Objects of this class can be directly visualized in Google Earth by using the plotKML-methods.

Slots

variable:

object of class "character"; variable name

sampled:

object of class "SpatialPoints"; one or more points that can be used to visualize temporal changes in the target variable

rasters:

object of class "RasterBrick"; a time-series of raster objects

TimeSpan.begin:

object of class "POSIXct"; begin of sampling for each raster map

TimeSpan.end:

object of class "POSIXct"; end of sampling for each raster map

Methods

plotKML

signature(obj = "RasterBrickTimeSeries"): plots time-series of rasters in Google Earth

Author(s)

Tomislav Hengl

See Also

RasterBrickSimulations-class, plotKML-methods

Examples


data(LST)
gridded(LST) <- ~x+y
proj4string(LST) <- CRS("+proj=utm +zone=33 +datum=WGS84 +units=m")
dates <- sapply(strsplit(names(LST), "LST"), function(x){x[[2]]})
datesf <- format(as.Date(dates, "%Y_%m_%d"), "%Y-%m-%dT%H:%M:%SZ")
# begin / end dates +/- 4 days:
TimeSpan.begin = as.POSIXct(unclass(as.POSIXct(datesf))-4*24*60*60, origin="1970-01-01") 
TimeSpan.end = as.POSIXct(unclass(as.POSIXct(datesf))+4*24*60*60, origin="1970-01-01")
LST_ll <- reproject(LST)
# pick few climatic stations:
data(HRtemp08)
pnts <- HRtemp08[which(HRtemp08$NAME=="Pazin")[1],]
pnts <- rbind(pnts, HRtemp08[which(HRtemp08$NAME=="Crni Lug - NP Risnjak")[1],])
pnts <- rbind(pnts, HRtemp08[which(HRtemp08$NAME=="Cres")[1],])
coordinates(pnts) <- ~Lon + Lat
proj4string(pnts) <- CRS("+proj=longlat +datum=WGS84")
## Not run: # get the dates from the file names:
LST <- brick(LST)
LST@title = "Time series of MODIS Land Surface Temperature (8-day mosaics) images"
LST.ts <- new("RasterBrickTimeSeries", variable = "LST", 
  sampled = pnts, rasters = LST_ll, TimeSpan.begin = TimeSpan.begin, TimeSpan.end = TimeSpan.end)
data(SAGA_pal)
# plot MODIS images in Google Earth:
plotKML(LST.ts, colour_scale=SAGA_pal[[1]])

## End(Not run)

[Package plotKML version 0.2-1 Index]