A class for outputs of analysis produced using the dismo package (MaxEnt)

A class containing input and output data produced by running the maxent (Maximum Entropy) species distribution modeling algorithm. Object of this class can be directly visualized in Google Earth by using the plotKML-method.

Slots

sciname:

object of class "character"; vector of species name compatible with the rgbif package; usually latin "genus" and "species" name

occurrences:

object of class "SpatialPoints"; occurrence-only records

TimeSpan.begin:

object of class "POSIXct"; begin of the sampling period

TimeSpan.end:

object of class "POSIXct"; end of the sampling period

maxent:

object of class "MaxEnt" (species distribution model); produced as an output of the dismo::maxent function or similar

sp.domain:

object of class "Spatial" (ideally "SpatialPolygonsDataFrame" or "SpatialPixelsDataFrame"); assumed spatial domain that can be set by the user or it will be estimated by MaxEnt (see examples below)

predicted:

object of class "RasterLayer"; contains results of prediction produced using the MaxEnt software

Methods

plotKML

signature(obj = "SpatialMaxEntOutput"): plots all MaxEnt output objects in Google Earth

Note

MaxEnt requires the maxent.jar file to be in the 'java' folder of the dismo package (see: system.file("java", package="dismo")). For more info refer to the dismo package documentation. Alternatively use the maxlike package (Royle et al. 2012), which does not require Java.

References

Hijmans, R.J, Elith, J., (2012) Species distribution modeling with R. CRAN, Vignette for the dismo package, 72 p.

Royle, J.A., Chandler, R.B., Yackulic, C. and J. D. Nichols. (2012) Likelihood analysis of species occurrence probability from presence-only data for modelling species distributions. Methods in Ecology and Evolution. doi:10.1111/j.2041-210X.2011.00182.x

dismo package (https://CRAN.R-project.org/package=dismo)

maxlike package (https://CRAN.R-project.org/package=maxlike)

rgbif package (https://CRAN.R-project.org/package=rgbif)

Author

Tomislav Hengl

See also