SpatialPredictions-class {plotKML}	R Documentation

A class for spatial predictions produced using gstat

Description

A class containing input and output maps generated through the process of geostatistical mapping. Object of this class can be directly visualized in Google Earth by using the plotKML-methods.

Slots

variable:

object of class "character"; variable name

observed:

object of class "SpatialPointsDataFrame" (must be 2D); see sp::SpatialPointsDataFrame

glm:

object of class "glm"; contains the summary of the regression model

vgmModel:

object of class "data.frame"; contains the variogram parameters passed from gstat

predicted:

object of class "SpatialPixelsDataFrame"; see sp::SpatialPixelsDataFrame

validation:

object of class "SpatialPointsDataFrame" containing results of validation

Methods

plotKML

signature(obj = "SpatialPredictions"): plots all objects in Google Earth

summary

signature(obj = "SpatialPredictions"): summarize object by showing the mapping accuracy (cross-validation) and the amount of variation explained by the model

Note

"SpatialPredictions" saves results of predictions for a single target variable, which can be of type numeric or factor. Multiple variables can be combined into a list.

Author(s)

Tomislav Hengl

See Also

fit.gstatModel, gstat::gstat-class, RasterBrickSimulations-class

Examples

 
# load observations:
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
# load grids:
data(meuse.grid)
coordinates(meuse.grid) <- ~x+y
gridded(meuse.grid) <- TRUE
proj4string(meuse.grid) <- CRS("+init=epsg:28992")
# fit a model:
omm <- fit.gstatModel(observations = meuse, formulaString = om~dist, covariates = meuse.grid)
show(omm@regModel)
# produce SpatialPredictions:
om.rk <- predict(omm, predictionLocations = meuse.grid)
# overview of the prediction process:
show(om.rk)
# plot the whole project in Google Earth:
## Not run: 
plotKML(om.rk)

## End(Not run)

[Package plotKML version 0.1-5 Index]