A class for spatial simulations containing equiprobable line, point or polygon features

A class containing input and output maps generated as equiprobable simulations of the same discrete object (for example multiple realizations of stream networks). Objects of this type can be directly visualized in Google Earth by using the plotKML-method.

Slots

realizations:

object of class "list"; multiple realizations of the same feature e.g. multiple realizations of stream network

summaries:

object of class "SpatialGridDataFrame"; summary measures

Methods

plotKML

signature(obj = "SpatialVectorsSimulations"): plots simulated vector objects and summaries (grids) in Google Earth

Author

Tomislav Hengl

See also

Examples

## load a list of equiprobable streams:
data(barstr)
data(bargrid)
library(sp)
coordinates(bargrid) <- ~ x+y
gridded(bargrid) <- TRUE
## output topology:
cell.size = bargrid@grid@cellsize[1]
bbox = bargrid@bbox
nrows = round(abs(diff(bbox[1,])/cell.size), 0) 
ncols = round(abs(diff(bbox[2,])/cell.size), 0)
gridT = GridTopology(cellcentre.offset=bbox[,1], cellsize=c(cell.size,cell.size), 
  cells.dim=c(nrows, ncols))
if (FALSE) ## derive summaries (observed frequency and the entropy or error):
bar_sum <- count.GridTopology(gridT, vectL=barstr[1:5]) 
## NOTE: this operation can be time consuming!
## plot the whole project and open in Google Earth:
plotKML(bar_sum, grid2poly = TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'obj' in selecting a method for function 'plotKML': object 'bar_sum' not found