Bigfoot reports (USA)

2984 observations of bigfoot (with attached dates). The field occurrence records have been obtained from the BigFoot Research Organization (BFRO) website. The BFRO reports generally consist of a description of the event and where it occurred, plus the quality classification. Similar data set has been used by Lozier et al. (2009) to demonstrate possible miss-interpretations of the results of species distribution modeling. The maps in the USAWgrids data set represent typical gridded environmental covariates used for species distribution modeling.

data(bigfoot)

Format

The bigfoot data frame contains the following columns:

Lon

a numeric vector; x-coordinate / longitude in the WGS84 system

Lat

a numeric vector; y-coordinate / latitude in the WGS84 system

NAME

name assigned by the observer (usually referent month / year)

DATE

'POSIXct' class vector

TYPE

confidence levels; according to the BFRO website: "Class A" reports involve clear sightings in circumstances where misinterpretation or misidentification of other animals can be ruled out with greater confidence; "Class B" and "Class C" reports are less credible.

The USAWgrids data frame (46,018 pixels; Washington, Oregon, Nevada and California state) contains the following columns:

globedem

a numeric vector; elevations from the ETOPO1 Global Relief Model

nlights03

an integer vector; lights at night image for 2003 (Version 2 DMSP-OLS Nighttime Lights Time Series)

sroads

a numeric vector; distance to main roads and railroads (National Atlas of the United States)

gcarb

a numeric vector; Global Biomass Carbon Map (New IPCC Tier-1 Global Biomass Carbon Map for the Year 2000)

dTRI

a numeric vector; density of pollutant releases (North American Pollutant Releases and Transfers database)

twi

a numeric vector; Topographic Wetness Index based on the globedem

states

an integer vector; USA states

globcov

land cover classes based on the MERIS FR images (GlobCover Land Cover version V2.2)

s1

a numeric vector; x-coordinates in the Albers equal-area projection system

s2

a numeric vector; y-coordinates in the Albers equal-area projection system

Author

Tomislav Hengl

References

Lozier, J.D., Aniello, P., Hickerson, M.J., (2009) Predicting the distribution of Sasquatch in western North America: anything goes with ecological niche modelling. Journal of Biogeography, 36(9):1623-1627. doi:10.1111/j.1365-2699.2009.02152.x

BigFoot Research Organization (http://www.bfro.net)

Note

According to the Time.com, a team of a dozen-plus experts from as far afield as Canada and Sweden have proclaimed themselves 95 percent certain of the mythical animal's existence on Kemerovo region territory some 3,000 kilometers east of Moscow (announced at the Tashtagol conference in 2011).

Examples

if (FALSE) # Load the BFRO records:
library(sp)
data(bigfoot)
aea.prj <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 
+x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"
library(sp)
coordinates(bigfoot) <- ~Lon+Lat
proj4string(bigfoot) <- CRS("+proj=latlon +datum=WGS84")
#> Error in CRS("+proj=latlon +datum=WGS84"): northings must follow eastings: +proj=latlon +datum=WGS84
library(rgdal)
bigfoot.aea <- spTransform(bigfoot, CRS(aea.prj))
#> Warning: NULL source CRS comment, falling back to PROJ string
#> Error in if (is.projected(obj)) {    tg <- wkt(obj)    if (is.null(tg)) {        tg <- proj4string(obj)        if (length(grep("+init", tg)) > 0L) {            tg <- wkt(CRS(tg))        }    }    o <- try(project(t(bbox(obj))[, 1:2], tg, inv = TRUE, use_aoi = FALSE),         silent = TRUE)    if (inherits(o, "try-error"))         return(NULL)} else {    o <- t(bbox(obj))[, 1:2]}: missing value where TRUE/FALSE needed
# Load the covariates:
data(USAWgrids)
gridded(USAWgrids) <- ~s1+s2
proj4string(USAWgrids) <- CRS(aea.prj)
# Visualize data:
data(SAGA_pal)
pnts <- list("sp.points", bigfoot.aea, pch="+", col="yellow")
#> Error in eval(expr, envir, enclos): object 'bigfoot.aea' not found
spplot(USAWgrids[2], col.regions=rev(SAGA_pal[[3]]), sp.layout=pnts)
#> Error in append(list(formula, data = as(sdf, "data.frame"), aspect = aspect,     panel = panel, xlab = xlab, ylab = ylab, scales = scales,     sp.layout = sp.layout, xlim = xlim, ylim = ylim, col.regions = col.regions),     list(...)): object 'pnts' not found