dailymeteo {plotKML}	R Documentation

Daily temperatures and precipitation for Croatia for year 2008

Description

The daily measurements of temperature (thermometers) and precipitation (rain gauges) for year 2008 kindly contributed by the Croatian National Meteorological Service. All data is projected in the UTM zone 33N system. HRtemp08 contains 56,608 measurements of temperature (159 stations by 365 days); HRprec08 contains 175,059 measurements of precipitation sums (489 stations by 365 days).

Usage

data(HRtemp08)

Format

The HRtemp08/HRprec08 data frames contain the following columns:

NAME

name of the meteorological station

Lon

a numeric vector; x-coordiante / longitude in the WGS84 system

Lat

a numeric vector; y-coordinate / latitude in the WGS84 system

DATE

'Date' class vector

TEMP/PREC

daily temperature measurements in degree C / cummulative precipitation in mm measured the day before

Note

The precision of the temperature readings in HRtemp08 is tenth of degree C. On most climatological stations temperature is measured three times a day, at 7 a.m., 1 p.m. and 9 p.m. The daily mean can be calculated as a weighted average. The precipitation estimates in mm (HRprec08) are collected in a bottle within the rain gauge and readings are usually manual by an observer at 7 a.m. The precipitation collected in the morning refer to the precipitation for previous 24 hours.

Author(s)

Tomislav Hengl and Melita Percec Tadic

References

Hengl, T., Heuvelink, G.B.M., Percec Tadic, M., Pebesma, E., (2011) Spatio-temporal prediction of daily temperatures using time-series of MODIS LST images. Theoretical and Applied Climatology.

http://spatial-analyst.net/book/HRclim2008

Examples


data(HRtemp08)
HRtemp08[1,]
p1 = newXMLNode("Placemark")
begin <- format(HRtemp08[1,"DATE"]-.5, "%Y-%m-%dT%H:%M:%SZ")
end <- format(HRtemp08[1,"DATE"]+.5, "%Y-%m-%dT%H:%M:%SZ")
txt <- sprintf('<name>%s</name><TimeStamp><begin>%s</begin><end>%s</end></TimeStamp><Point><coordinates>%.4f,%.4f,%.0f</coordinates></Point>', HRtemp08[1,"NAME"], begin, end, HRtemp08[1,"Lon"], HRtemp08[1,"Lat"], 0)
parseXMLAndAdd(txt, parent=p1)
p1
## Not run: 
## Load precipitation data:
load(url("http://plotkml.r-forge.r-project.org/HRprec08.rda"))
## Plot measured temperatures and precipitation in Google Earth:

## End(Not run)

[Package plotKML version 0.1-2 Index]