eberg {plotKML}	R Documentation

Ebergotzen – Digital Soil Mapping case study

Description

Ebergötzen is 10 by 10 km study area in the vicinity of the city of Göttingen in Central Germany. This area has been extensively surveyed over the years, mainly for the purposes of developing operational digital soil mapping techniques (Gehrt and Böhner, 2001).
eberg table contains 2937 observations (augers) of soil textures and soil types. eberg_grid contains three gridded maps that can be used as covariates for spatial prediction of soil variables. eberg_zones is a polygon map showing the distribution of parent material (Silt and sand, Sandy material, Clayey derivats, Clay and loess). eberg_contours shows contours lines derived using the 25~m DEM of the area with 10~m equidistance.

Usage

data(eberg)

Format

The eberg data frame contains the following columns:

ID

universal identifier

X

a numeric vector; x-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

Y

a numeric vector; y-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

SAND

topsoil sand content estimated by hand; see Ad-hoc-AG Boden (2005)

SILT

topsoil silt content estimated by hand

CLAY

topsoil clay content estimated by hand

SOILTYPE

a vector containing factors; soil classes according to the German soil classification system (13 classes in this data set)

The eberg_grid is SpatialGridDataFrame that contains the following layers:

DEM

elevation values from the topographic map

TWI

Topographic Wetness Index derived using the SAGA algorithm

Z

containing factors; four soil topographic/geological units (soil mapping units)

Note

Texture by hand method (Ad-hoc-AG Boden, 2005) in eberg can be used to determine the content of soil earth fractions only to an accuracy of ±5-10% (Skaggs et al. 2001). A surveyor distinguishes to which of the 32 texture classes a soil samples belongs to, and then estimates the content of fractions. E.g. texture class St2 has 10% clay, 25% silt and 65% sand.

Author(s)

The Ebergötzen dataset is courtesy of Gehrt Ernst (Ernst.Gehrt@niedersachsen.de), the State Authority for Mining, Energy and Geology, Hannover, Germany and Olaf Conrad, University of Hamburg (conrad@geowiss.uni-hamburg.de). The data set has been prepared for this exercise by Tomislav Hengl (tom.hengl@wur.nl).

References

Ad-hoc-AG Boden, (2005) Bodenkundliche Kartieranleitung. 5th Ed, Bundesanstalt fur Geowissenschaften und Rohstoffe und Niedersächsisches Landesamt für Bodenforshung, Hannover, p. 423.

Böhner, J., McCloy, K. R. and Strobl, J. (Eds), (2006) SAGA — Analysis and Modelling Applications. Göttinger Geographische Abhandlungen, Heft 115. Verlag Erich Goltze GmbH, Göttingen, 117 pp.

Gehrt, E., Böhner, J., (2001) Vom punkt zur flache — probleme des ‘upscaling’ in der bodenkartierung. In: Diskussionsforum Bodenwissenschaften: Vom Bohrstock zum Bildschirm. FH, Osnabrück, pp. 17-34.

Skaggs, T. H., Arya, L. M., Shouse, P. J., Mohanty, B. P., (2001) Estimating Particle-Size Distribution from Limited Soil Texture Data. Soil Science Society of America Journal 65 (4): 1038-1044.

http://geomorphometry.org/content/ebergotzen

Examples


data(eberg)
data(eberg_grid)
data(eberg_zones)
data(eberg_contours)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
## Not run: 
l.sp <- list("sp.lines", eberg_contours, col="black")
spplot(eberg_grid["DEM"], col.regions = SAGA_pal[[1]], sp.layout=l.sp)
spplot(eberg_zones, sp.layout=list("sp.points", eberg, col="black", pch="+"))
## End(Not run)

[Package plotKML version 0.1-2 Index]