eberg {plotKML}	R Documentation

Ebergotzen — soil mapping case study

Description

Ebergötzen is 10 by 10 km study area in the vicinity of the city of Göttingen in Central Germany. This area has been extensively surveyed over the years, mainly for the purposes of developing operational digital soil mapping techniques (Gehrt and Böhner, 2001), and has been used by the SAGA GIS development team to demonstrate various processing steps.
eberg table contains 3670 observations (augers) of soil textures at five depths (0-10, 10-30, 30-50, 50-70, and 70-90), and field records of soil types according to the German soil classification system. eberg_grid contains gridded maps at 100 m resolution that can be used as covariates for spatial prediction of soil variables. eberg_grid25 contains grids at finer resolution (25 m). eberg_zones is a polygon map showing the distribution of parent material (Silt and sand, Sandy material, Clayey derivats, Clay and loess). eberg_contours shows contour lines derived from the 25 m DEM of the area using 10 m equidistance.

Usage

data(eberg)

Format

The eberg data frame (irregular points) contains the following columns:

ID

universal identifier

soiltype

a vector containing factors; soil classes according to the German soil classification system: "A" (Auenboden), "B" (Braunerde), "D" (Pelosol), "G" (Gley), "Ha" (Moor), "Hw" (HMoor), "K" (Kolluvisol), "L" (Parabraunerde), "N" (Ranker), "Q" (Regosol), "R" (Rendzina), "S" (Pseudogley), "Z" (Pararendzina)

TAXGRSC

a vector containing factors; full soil class names according to the German soil classification system (see soiltype column)

X

a numeric vector; x-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

Y

a numeric vector; y-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

UHDICM_*

a numeric vector; upper horizon depth in cm per horizon

LHDICM_*

a numeric vector; lower horizon depth in cm per horizon

SNDMHT_*

a numeric vector; sand content estimated by hand per horizon (0-100 percent); see Ad-hoc-AG Boden (2005) for more details

SLTMHT_*

a numeric vector; silt content estimated by hand per horizon (0-100 percent)

CLYMHT_*

a numeric vector; clay content estimated by hand per horizon (0-100 percent)

The eberg_grid data frame (regular grid at 100 m resolution) contains the following columns:

PRMGEO6

a vector containing factors, parent material classes from the geological map (mapping units)

DEMSRT6

a numeric vector; elevation values from the SRTM DEM

TWISRT6

a numeric vector; Topographic Wetness Index derived using the SAGA algorithm

TIRAST6

a numeric vector; Thermal Infrared (TIR) reflection values from the ASTER L1 image band 14 (2010-06-05T10:26:50Z) obtained via the NASA's GloVis browser

LNCCOR6

a vector containing factors; Corine Land Cover 2006 classes

x

a numeric vector; x-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

y

a numeric vector; y-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

The eberg_grid25 data frame (regular grid at 25 m resolution) contains the following columns:

DEMTOPx

a numeric vector; elevation values from the topographic map

HBTSOLx

a vector containing factors; main soil type according to the German soil classification system (see column "soiltype" above) estimated per crop field

TWITOPx

a numeric vector; Topographic Wetness Index derived using the SAGA algorithm

NVILANx

a numeric vector; NDVI image derived using the Landsat image from the Image 2000 project

x

a numeric vector; x-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

y

a numeric vector; y-coordinate (m) in DHDN / Gauss-Krueger zone 3 (German coordinate system)

Note

Texture by hand method can be used to determine the content of soil earth fractions only to an accuracy of ±5-10% (Skaggs et al. 2001). A surveyor distinguishes to which of the 32 texture classes a soil samples belongs to, and then estimates the content of fractions; e.g. texture class St2 has 10% clay, 25% silt and 65% sand (Ad-hoc-AG Boden, 2005).

Author(s)

The Ebergötzen dataset is courtesy of Gehrt Ernst (Ernst.Gehrt@niedersachsen.de), the State Authority for Mining, Energy and Geology, Hannover, Germany and Olaf Conrad, University of Hamburg (conrad@geowiss.uni-hamburg.de). The original data set has been prepared for this exercise by Tomislav Hengl (tom.hengl@wur.nl).

References

Ad-hoc-AG Boden, (2005) Bodenkundliche Kartieranleitung. 5th Ed, Bundesanstalt fur Geowissenschaften und Rohstoffe und Niedersächsisches Landesamt für Bodenforshung, Hannover, p. 423.

Böhner, J., McCloy, K. R. and Strobl, J. (Eds), (2006) SAGA — Analysis and Modelling Applications. Göttinger Geographische Abhandlungen, Heft 115. Verlag Erich Goltze GmbH, Göttingen, 117 pp.

Gehrt, E., Böhner, J., (2001) Vom punkt zur flache — probleme des ‘upscaling’ in der bodenkartierung. In: Diskussionsforum Bodenwissenschaften: Vom Bohrstock zum Bildschirm. FH, Osnabrück, pp. 17-34.

Skaggs, T. H., Arya, L. M., Shouse, P. J., Mohanty, B. P., (2001) Estimating Particle-Size Distribution from Limited Soil Texture Data. Soil Science Society of America Journal 65 (4): 1038-1044.

http://geomorphometry.org/content/ebergotzen

Examples


data(eberg)
data(eberg_grid)
data(eberg_zones)
data(eberg_contours)
library(sp)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
gridded(eberg_grid) <- ~x+y
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
# visualize the maps:
data(SAGA_pal)
l.sp <- list("sp.lines", eberg_contours, col="black")
## Not run: 
spplot(eberg_grid["DEMSRT6"], col.regions = SAGA_pal[[1]], sp.layout=l.sp)
spplot(eberg_zones, sp.layout=list("sp.points", eberg, col="black", pch="+"))

## End(Not run)

[Package plotKML version 0.4-1 Index]