kml {plotKML}	R Documentation

Wrapper function to plot any Spatial* or Raster* object in a KML file.

Description

This function allows the user to plot any Spatial* object (from the sp package) or Raster* object (from the raster package) to be plotted in a KML file. Various aesthetics, i.e. ways to represent information, are provided (colour, transparency, size, shape, etc.). Their availability depends on the class of the object to plot.

Usage

kml(obj, ...)

Arguments

obj 	 An object inheriting from the Spatial or the BasicRaster classes.

... 	 Arguments interpreted in the context of kml() (see Details below)

Value

A KML file located as per the file option.

Author(s)

Pierre Roudier, Tomislav Hengl, Dylan Beaudette

See Also

kml_open, kml_close, kml_layer

Examples


# Plotting a SpatialPointsDataFrame object
data(eberg)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
## Not run: 
# Simple plot
kml(eberg, file = "eberg-0.kml")
# Plot using aesthetics
kml(eberg, colour = SAND, size = CLAY, alpha = 0.75, file = "eberg-1.kml")

## End(Not run)
# Plotting a SpatialPixelsDataFrame or a Raster* object
data(eberg_grid)
## Not run: 
kml(eberg_grid, colour = DEM, file = "eberg_grid-0.kml")

## End(Not run)

[Package plotKML version 0.1-2 Index]