kml_metadata-methods {plotKML}	R Documentation

Add metadata table to the active layer

Description

Adds a selection of metadata to the description box of an active layer.

Usage


## S4 method for signature 'SpatialMetadata'
kml_metadata(obj, cwidth = 150, twidth = 500, asText = FALSE)

Arguments

obj
object of class "SpatialMetadata"

cwidth
html column width for the field names

twidth
html total table width

asText
logical; return the output as XML or characters

Details

The kml_metadata function, by default, prints out only a number of selected metadata fields:

"Citation_title",

"Abstract",

"Object_Count",

"Beginning_Date",

"Ending_Date",

"Data_Order_URL",

"Other_Citation_Details",

"Citation_URL",

"Data_Set_Credit",

"Data_Distributing_Organization",

"Format_Information_Content",

"Native_Data_Set_Environment"

See data(mdnames) for a complete list of metadata fields.

Author(s)

Tomislav Hengl

See Also

spMetadata

[Package plotKML version 0.5-6 Index]