kml_open {plotKML}	R Documentation

Opens / closes a KML file connection

Description

Opens a KML file in write mode and initiates the KML header. The same file connection is further accessible by other kml_*() functions such as kml_layer() and kml_close(). This allows creation of customized multi-layered KML files.

Usage


kml_open(file, name = file, overwrite = FALSE, kml.url = get("kml_url", envir = plotKML.opts))

Arguments

file 	 either a character string naming a file.

name 	 A name for the KML file to be created.

overwrite 	 Logical. If TRUE, "name" will be overwritten if it exists.

kml.url 	 URL of the KML standard to be used.

Author(s)

Pierre Roudier and Dylan Beaudette

See Also

kml, kml_layer

[Package plotKML version 0.1-2 Index]