kml_layer.Raster {plotKML}	R Documentation

Writes raster objects to KML

Description

Writes rasters to PNG images and makes a KML code (ground overlays). Works with "RasterLayer" and "RasterStack" class objects. Target attributes can be specified using aesthetics arguments (e.g. "colour").

Usage


kml_layer.Raster(obj, plot.legend = TRUE, metadata = NULL, raster_name, 
     png.width = ncol(obj), png.height = nrow(obj), ...)

Arguments

obj
object of class "RasterLayer", "SpatialPixelsDataFrame" or "SpatialGridDataFrame"

plot.legend
logical; specify whether a map legend should be generated automatically

metadata
(optional) specify the metadata object

raster_name
(optional) specify the output file name (PNG)

png.width
(optional) width of the PNG file

png.height
(optional) height of the PNG file

...
additional aesthetics arguments

Details

Google Earth does not properly handle a 24-bit PNG file which has a single transparent color (read more at Google Earth Help). To force transparency, plotKML will try to set it using the -matte -transparent "\#FFFFFF" option in the ImageMagick convert program (ImageMagick needs to be installed separately and located using plotKML.env()).

Author(s)

Tomislav Hengl, Pierre Roudier and Dylan Beaudette

See Also

kml-methods, kml_open, kml_layer.RasterBrick, plotKML-method

Examples


data(eberg_grid)
library(sp)
coordinates(eberg_grid) <- ~x+y
gridded(eberg_grid) <- TRUE
proj4string(eberg_grid) <- CRS("+init=epsg:31467")
data(SAGA_pal)
library(raster)
r <- raster(eberg_grid["TWISRT6"])
## Not run: # KML plot with a single raster:
kml(r, colour_scale = SAGA_pal[[1]], colour = TWISRT6) 

## End(Not run)

[Package plotKML version 0.3-4 Index]