kml_layer.RasterBrick {plotKML}	R Documentation

Exports a time series of images to KML

Description

Writes a series of images to png and creates ground overlays. Works only with "RasterBrick" class objects with time dimension specified via the "@zvalue".

Usage


kml_layer.RasterBrick(obj, dtime = "", extrude = FALSE, tz = "GMT", z.lim = c("min", "max"),
   colour_scale = get("colour_scale_numeric", envir = plotKML.opts),
   spMetadata = NULL, html.table = NULL,  ...)

Arguments

obj 	 object of class "RasterBrick" (a time series of images)

dtime 	 temporal support (point or block) expressed in seconds

extrude 	 logical

tz = "GMT" 	 referent time zone

z.lim 	 upper and lower limits (unique for all maps in the time series); by default – the absolute minimum and maximum in values

colour_scale 	 color palette; by default using the color scale for numeric variables

spMetadata 	 (optional) specify the metadata object

html.table 	 optional description block (html) for each raster layer

... 	 additional style arguments (see aesthetics)

Details

This method is recommended for visualization of numeric bands representing the same variable i.e. time series of images. To export a stack of images of different type see kml_layer.Raster. If the "@zvalue" slot is empty, dates will be added by subtracting days from the current day with 1–day increments.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl)

See Also

kml, kml_open, kml_layer.Raster

Examples


data(LST)
data(SAGA_pal)
# time dimension defined via zvalue slot:
LST@zvalue[c(1,12,24)]
## Not run: 
# subset to 3 images:
LSTs <- brick(raster(LST, 1), raster(LST, 12), raster(LST, 24))
LST_ll <- reproject(LSTs, program="FWTools")
# copy the dates:
LST_ll@zvalue <- LST@zvalue[c(1,12,24)]
# KML plot with block temporal support of 8 days:
kml(LST_ll, colour_scale=SAGA_pal[[1]], dtime=8*24*60^2)
## End(Not run)

[Package plotKML version 0.1-2 Index]