kml_layer.RasterBrick {plotKML}	R Documentation

Export a time series of images to KML

Description

Writes a series of images to PNGs and uses them to create ground overlays. Works only with "RasterBrick" class objects with time dimension specified via the "@zvalue".

Usage


kml_layer.RasterBrick(obj, plot.legend = TRUE, dtime = "", tz = "GMT", 
          z.lim = c(min(minValue(obj), na.rm=TRUE), max(maxValue(obj), na.rm=TRUE)), 
          colour_scale = get("colour_scale_numeric", envir = plotKML.opts),
          home_url = get("home_url", envir = plotKML.opts),
          metadata = NULL, html.table = NULL,
          altitudeMode = "clampToGround", balloon = FALSE,
          png.width, png.height, min.png.width = 800, ...)

Arguments

obj
object of class "RasterBrick" (e.g. a time series of images)

plot.legend
logical; specify whether a map legend should be generated automatically

dtime
temporal support (point or block) expressed in seconds

tz
referent time zone

z.lim
upper and lower limits (unique for all maps in the time series); the function by default uses the absolute minimum and maximum in values

colour_scale
color palette; by default uses the color scale for numeric variables

home_url
(optional) URL directory / location of the images

metadata
(optional) the metadata object

html.table
(optional) the description block (html)

altitudeMode
character; the default altitudeMode

balloon
logical; specifies whether to display balloon for each element

png.width
(optional) width of the PNG files

png.height
(optional) height of the PNG files

min.png.width
(optional) minimum width of the PNG file

...
additional arguments (see aesthetics)

Details

This method is recommended for visualization of numeric bands representing the same variable i.e. time series of images. To export a stack of images of different type see kml_layer.Raster. If the "@zvalue" slot is empty, dates will be added by subtracting days from the current day with 1–day increments.

Author(s)

Tomislav Hengl

See Also

kml-methods, kml_open, kml_layer.Raster, plotKML-method

[Package plotKML version 0.4-5 Index]