Writes spatial lines to KML

Writes object of class "SpatialLines*" to KML with a possibility to parse attribute variables using several aesthetics arguments.

kml_layer.SpatialLines(obj, subfolder.name = paste(class(obj)), 
  extrude = FALSE, z.scale = 1, metadata = NULL, 
  html.table = NULL, TimeSpan.begin = "", TimeSpan.end = "", ...)

Arguments

obj

object of class "SpatialLines*"

subfolder.name

character; optional subfolder name

extrude

logical; specifies whether to connect the LinearRing to the ground

z.scale

vertical exaggeration

metadata

(optional) specify the metadata object

html.table

optional description block (html) for each GPS point (vertices)

TimeSpan.begin

(optional) beginning of the referent time period

TimeSpan.end

(optional) end of the referent time period

...

additional style arguments (see aesthetics)

Details

Only colour and width (aesthetics) are recommended when visualizing SpatialLines* objects.
TimeSpan.begin and TimeSpan.end are optional TimeStamp vectors in the format:
yyyy-mm-ddThh:mm:sszzzzzz
Use the same time values for both TimeSpan.begin and TimeSpan.end if the measurements refer to a single moment in time. TimeSpan.begin and TimeSpan.end can be either a single value or a vector of values.

Author

Pierre Roudier, Tomislav Hengl and Dylan Beaudette

See also

Examples

# \donttest{
library(rgdal)
library(sp)
data(eberg_contours)
data(SAGA_pal)
names(eberg_contours)
#> [1] "ID" "Z" 
# KML plot with elevations used as 'colour' argument:
kml.file = paste0(tempdir(), "/eberg_contours.kml")
kml(eberg_contours, file.name = kml.file, 
  colour_scale = SAGA_pal[[1]], colour = Z)
#> KML file opened for writing...
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html
#> Reprojecting to +proj=longlat +datum=WGS84 +no_defs ...
#> Writing to KML...
#> Closing  C:\Users\shs\AppData\Local\Temp\RtmpiQO1Iy/eberg_contours.kml
# }