Writes spatial points to KML

Writes object of class "SpatialPoints*" to KML with a possibility to parse attribute variables using several aesthetics arguments.

kml_layer.SpatialPoints(obj, subfolder.name = paste(class(obj)), 
        extrude = TRUE, z.scale = 1, 
        LabelScale = get("LabelScale", envir = plotKML.opts), 
        metadata = NULL, html.table = NULL, TimeSpan.begin = "",
        TimeSpan.end = "", points_names, ...)

Arguments

obj

object of class "SpatialPoints*"

subfolder.name

character; optional subfolder name

extrude

logical; specifies whether to connect the point to the ground with a line

z.scale

numeric; exaggeration in vertical dimension

LabelScale

numeric; scale factor for size of labels

metadata

(optional) specify the metadata object

html.table

(optional) specify the description block (html) for each point

TimeSpan.begin

(optional) beginning of the referent time period

TimeSpan.end

(optional) end of the referent time period

points_names

character; forces the point labels (size of the character vector must equal the number of the points)

...

additional style arguments (see aesthetics)

Details

TimeSpan.begin and TimeSpan.end are optional TimeStamp vectors:

yyyy-mm-ddThh:mm:sszzzzzz

For observations at point support (a single moment in time), use the same time values for both TimeSpan.begin and TimeSpan.end. TimeSpan.begin and TimeSpan.end can be either a single value or a vector of values.
Optional aesthetics arguments are shapes (icons), colour, sizes, altitude (if not a 3D object; variable to be used to specify altitude above ground), altitudeMode (altitude mode type (clampToGround, relativeToGround or absolute). Although this function can be used to plot over five variables, more than three aesthetics arguments is not recommended (e.g. limit to size and colour).

Author

Pierre Roudier, Tomislav Hengl and Dylan Beaudette

See also

Examples

data(eberg)
data(SAGA_pal)
library(sp)
library(rgdal)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
names(eberg)
#>  [1] "ID"       "soiltype" "TAXGRSC"  "UHDICM_A" "LHDICM_A" "SNDMHT_A"
#>  [7] "SLTMHT_A" "CLYMHT_A" "UHDICM_B" "LHDICM_B" "SNDMHT_B" "SLTMHT_B"
#> [13] "CLYMHT_B" "UHDICM_C" "LHDICM_C" "SNDMHT_C" "SLTMHT_C" "CLYMHT_C"
#> [19] "UHDICM_D" "LHDICM_D" "SNDMHT_D" "SLTMHT_D" "CLYMHT_D" "UHDICM_E"
#> [25] "LHDICM_E" "SNDMHT_E" "SLTMHT_E" "CLYMHT_E"
# subset to 10 percent:
eberg <- eberg[runif(nrow(eberg))<.1,]
if (FALSE) # plot the measured CLAY content:
kml(eberg, labels = CLYMHT_A)
shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png"
# color only:
kml(eberg, shape = shape, colour = SLTMHT_A, labels = "", colour_scale = SAGA_pal[[1]])
#> KML file opened for writing...
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html
#> Reprojecting to +proj=longlat +datum=WGS84 +no_defs ...
#> Writing to KML...
#> Closing  eberg.kml
# two variables at the same time:
kml(eberg, shape = shape, size = CLYMHT_A, colour = SLTMHT_A, labels = "")
#> KML file opened for writing...
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html
#> Reprojecting to +proj=longlat +datum=WGS84 +no_defs ...
#> Writing to KML...
#> Closing  eberg.kml
# two aesthetics elements are effective in emphasizing hot-spots:
kml(eberg, shape = shape, altitude = CLYMHT_A*10, extrude = TRUE, 
  colour = CLYMHT_A, labels = CLYMHT_A, kmz = TRUE)
#> KML file opened for writing...
#> Warning: CRS object has comment, which is lost in output; in tests, see
#> https://cran.r-project.org/web/packages/sp/vignettes/CRS_warnings.html
#> Reprojecting to +proj=longlat +datum=WGS84 +no_defs ...
#> Writing to KML...
#> Closing  eberg.kml


## example of how plotKML is programmed:
data(HRtemp08)
HRtemp08[1,]
#>       NAME      Lat      Lon       DATE TEMP
#> 1 Bilogora 45.88512 17.20566 2008-02-18 2.35
library(XML)
p1 = newXMLNode("Placemark")
begin <- format(HRtemp08[1,"DATE"]-.5, "%Y-%m-%dT%H:%M:%SZ")
end <- format(HRtemp08[1,"DATE"]+.5, "%Y-%m-%dT%H:%M:%SZ")
txt <- sprintf('<name>%s</name><TimeStamp><begin>%s</begin><end>%s</end></TimeStamp>
    <Point><coordinates>%.4f,%.4f,%.0f</coordinates></Point>', HRtemp08[1,"NAME"], 
    begin, end, HRtemp08[1,"Lon"], HRtemp08[1,"Lat"], 0)
parseXMLAndAdd(txt, parent=p1)
p1
#> <Placemark>
#>   <name>Bilogora</name>
#>   <TimeStamp>
#>     <begin>2008-02-17T00:00:00Z</begin>
#>     <end>2008-02-18T00:00:00Z</end>
#>   </TimeStamp>
#>   <Point>
#>     <coordinates>17.2057,45.8851,0</coordinates>
#>   </Point>
#> </Placemark>