kml_layer.SpatialPoints {plotKML}	R Documentation

Exports objects of type SpatialPoints* to KML

Description

Writes object of type SpatialPoints* to KML with several aesthetics arguments.

Usage


kml_layer.SpatialPoints(obj, extrude = FALSE, 
    LabelScale = get("LabelScale", envir = plotKML.opts), z.scale = 1, 
    metadata = NULL, html.table = NULL, ...)

Arguments

obj 	 object of class "RasterBrick" (a time series of images)

extrude 	 logical; specifies whether to connect the point to the ground with a line

LabelScale 	 numeric; default size of the labels

metadata 	 (optional) specify the metadata object

html.table 	 (optional) specify the description block (html) for each point

... 	 additional style arguments (see aesthetics)

labels 	 variable to be used to attach labels to spatial lines

shapes 	 variable to be used to attach icons

colour 	 variable to be used to color spatial lines

sizes 	 variable to be used control size of icons

altitude 	 if not a 3D object; variable to be used to specify altitude above ground

altitudeMode 	 altitude mode type (clampToGround, relativeToGround or absolute)

Details

Although this function can be used to plot over five variables, it is highly recommended that you use a maximum of two or three aesthetics arguments (e.g. size and colour).

Author(s)

Pierre Roudier (pierre.roudier@landcare.nz), Tomislav Hengl (tom.hengl@wur.nl), Dylan Beaudette (debeaudette@ucdavis.edu)

See Also

kml, kml_open

Examples


data(eberg)
data(SAGA_pal)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
names(eberg)
## Not run: 
# subset:
eberg <- eberg[runif(nrow(eberg))<.2,]
# KML plot measured CLAY content:
kml(eberg, labels=CLAY)
shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png"
kml(eberg, shape = shape, colour = SILT, labels = "", colour_scale = SAGA_pal[[1]])
# Two variables at the same time:
kml(eberg, shape = shape, size = CLAY, colour = SILT, labels = "")
# Elevation used for visualization 
# two aesthetics elements emphasize hot-spots:
kml(eberg, shape = shape, altitude = CLAY*10, extrude = TRUE, colour = CLAY, labels = CLAY)
## End(Not run)

[Package plotKML version 0.1-2 Index]