kml_layer.SpatialPoints {plotKML}	R Documentation

Export spatial points to KML

Description

Writes object of class "SpatialPoints*" to KML with a possibility to parse attribute variables using several aesthetics arguments.

Usage


kml_layer.SpatialPoints(obj, extrude = FALSE, 
    LabelScale = get("LabelScale", envir = plotKML.opts), z.scale = 1, 
    metadata = NULL, html.table = NULL, TimeSpan.begin = "", TimeSpan.end = "", ...)

Arguments

obj 	 object of class "SpatialPoints*"

extrude 	 logical; specifies whether to connect the point to the ground with a line

LabelScale 	 numeric; default size of the labels

metadata 	 (optional) specify the metadata object

html.table 	 (optional) specify the description block (html) for each point

TimeSpan.begin 	 (optional) beginning of the referent time period

TimeSpan.end 	 (optional) end of the referent time period

... 	 additional style arguments (see aesthetics)

labels 	 variable to be used to attach labels to spatial lines

shapes 	 variable to be used to attach icons

colour 	 variable to be used to color spatial lines

sizes 	 variable to be used control size of icons

altitude 	 if not a 3D object; variable to be used to specify altitude above ground

altitudeMode 	 altitude mode type (clampToGround, relativeToGround or absolute)

Details

TimeSpan.begin and TimeSpan.end are optional TimeStamp vectors (yyyy-mm-ddThh:mm:sszzzzzz). For observations at point support (a single moment in time), use the same time values for both TimeSpan.begin and TimeSpan.end. TimeSpan.begin and TimeSpan.end can be either a single value or a vector of values.
Although this function can be used to plot over five variables, more than three aesthetics arguments is not recommended (e.g. limit to size and colour).

Author(s)

Pierre Roudier, Tomislav Hengl and Dylan Beaudette

See Also

kml_layer.STIDFtraj

Examples


data(eberg)
data(SAGA_pal)
coordinates(eberg) <- ~X+Y
proj4string(eberg) <- CRS("+init=epsg:31467")
names(eberg)
## Not run: 
# subset to 20 percent:
eberg <- eberg[runif(nrow(eberg))<.2,]
# KML plot measured CLAY content:
kml(eberg, labels = CLAY)
shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png"
kml(eberg, shape = shape, colour = SILT, labels = "", colour_scale = SAGA_pal[[1]])
# Two variables at the same time:
kml(eberg, shape = shape, size = CLAY, colour = SILT, labels = "")
# Two aesthetics elements are effective in emphasizing hot-spots:
kml(eberg, shape = shape, altitude = CLAY*10, extrude = TRUE, colour = CLAY, labels = CLAY, kmz = TRUE)

## End(Not run)

[Package plotKML version 0.1-2 Index]