kml_layer.SpatialPolygons {plotKML}	R Documentation

Exports objects of type SpatialPolygons* to KML

Description

Writes object of type SpatialPolygons* to KML with several aesthetics arguments.

Usage


kml_layer.SpatialPolygons(obj, extrude = FALSE, tessellate = FALSE, outline = TRUE, 
   plot.labpt = FALSE, z.scale = 1, LabelScale = get("LabelScale", envir = plotKML.opts),
   metadata = NULL, html.table = NULL, ...)

Arguments

obj 	 object of class "RasterBrick" (a time series of images)

extrude 	 logical; specifies whether to connect the point to the ground with a line

tessellate 	 logical; specifies whether to connect the LinearRing to the ground

outline 	 logical; specifies whether to outline the polygon

plot.labpt 	 logical; specifies whether to add the label point (polygon centre)

spMetadata 	 (optional) specify the metadata object

html.table 	 optional description block (html) for each GPS point (vertices)

... 	 additional style arguments (see aesthetics)

labels 	 variable to be used to attach labels to polygon centres

colour 	 variable to be used to color polygons

alpha 	 variable to be used to set transparency

altitude 	 if not a 3D object; variable to be used to specify altitude above ground

altitudeMode 	 altitude mode type ("clampToGround", "relativeToGround" or "absolute")

Details

Label points are be default not plotted (we recommend adding the legend to attribute maps instead). Transparency can be set by using the alpha argument.

Author(s)

Pierre Roudier (pierre.roudier@landcare.nz), Tomislav Hengl (tom.hengl@wur.nl), Dylan Beaudette (debeaudette@ucdavis.edu)

See Also

kml, kml_open, kml_layer.SpatialLines

Examples


data(eberg_zones)
names(eberg_zones)
## Not run: 
# Random colors:
kml(eberg_zones)
kml(eberg_zones, plot.labpt = TRUE, labels = ZONES)
## End(Not run)

[Package plotKML version 0.1-2 Index]