metadata2SLD-methods {plotKML}	R Documentation

Methods to create a Styled Layer Description (SLD) file

Description

Creates a Styled Layer Description (SLD) file, that can be attached to a spatial layer contributed to GeoServer. It writes the "sp.pallete" object (legend entries, titles and colors) to an external file.

Usage


## S4 method for signature 'SpatialMetadata'
metadata2SLD(obj, ...)

Arguments

obj
object of class "SpatialMetadata"

...
other arguments

Details

The structure of the SLD file is determined by the object class (Point, Polygon, SpatialPixels).

Author(s)

Tomislav Hengl

See Also

metadata2SLD.SpatialPixels, spMetadata

[Package plotKML version 0.5-0 Index]