plotKML-package {plotKML}	R Documentation

Plotting of space-time objects in Google Earth

Description

A suite of functions for converting sp-class objects into KML or KMZ documents for use in Google Earth.

Details

Package: 	 plotKML

Type: 	 Package

Version: 	 0.1

Date: 	 2011-05-25

License: 	 GPL

LazyLoad: 	 yes

Note

This package has been developed as a part of the Global Soil Information Facilities project, which is run jointly by the ISRIC Institute and collaborators.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl), Pierre Roudier (pierre.roudier@landcare.nz), Dylan Beaudette (debeaudette@ucdavis.edu)

References

Package home URL (http://plotkml.r-forge.r-project.org/)

See also: Google Earth Outreach project (http://earth.google.com/outreach/tutorials.html)

[Package plotKML version 0.1-2 Index]