plotKML-package {plotKML}	R Documentation

Visualization of spatial and spatio-temporal objects in Google Earth

Description

A suite of functions for converting 2D and 3D spatio-temporal (sp, raster and spacetime package classes) objects into KML or KMZ documents for use in Google Earth.

Details

Package: 	 plotKML

Type: 	 Package

URL: 	 http://plotkml.r-forge.r-project.org/

License: 	 GPL

LazyLoad: 	 yes

Note

This package has been developed as a part of the Global Soil Information Facilities project, which is run jointly by the ISRIC Institute and collaborators. ISRIC is a non-profit organization with a mandate to serve the international community as custodian of global soil information and to increase awareness and understanding of the role of soils in major global issues.

Author(s)

Tomislav Hengl (tom.hengl@wur.nl), Pierre Roudier (pierre.roudier@landcare.nz), Dylan Beaudette (debeaudette@ucdavis.edu), Daniel Nüst (d.nuest@52north.org)

References

KML documentation (http://code.google.com/apis/kml/documentation/)

Google Earth Outreach project (http://earth.google.com/outreach/tutorials.html)

[Package plotKML version 0.2-4 Index]