plotKML specific environmental variables / paths

Sets the environmental, package specific parameters and settings (URLs, names, default color palettes and similar) that can be later on passed to other functions.

plotKML.env(colour_scale_numeric = "", colour_scale_factor = "",
        colour_scale_svar = "", ref_CRS, NAflag, icon, LabelScale, size_range, 
        license_url, metadata_sel, kmz, kml_xsd, kml_url, kml_gx, gpx_xsd,
        fgdc_xsd, inspire_xsd, convert, gdalwarp, gdal_translate, python,
        home_url, show.env = TRUE, silent = TRUE)

Arguments

colour_scale_numeric

default colour scheme for numeric variables

colour_scale_factor

default colour scheme for factor variables

colour_scale_svar

default colour scheme for model error (e.g. mapping error)

ref_CRS

the referent CRS proj4string ("+proj=longlat +datum=WGS84")

NAflag

the default missing value flag (usually "-99999")

icon

the default icon URL

LabelScale

the default scale factor for labels

size_range

the default size range

license_url

the default license URL

metadata_sel

a list of the default metadata fields for summary

kmz

logical; the default compression setting

kml_xsd

the default KML scheme URL

kml_url

the default KML format URL

kml_gx

the default extended KML scheme URL

gpx_xsd

the default GPX scheme URL

fgdc_xsd

the default metadata scheme URL

inspire_xsd

the default metadata scheme URL

convert

a path to ImageMagick convert program

gdalwarp

a path to gdalwarp program

gdal_translate

a path to gdalwarp program

python

a path to Python program

home_url

the default location of all icons and auxiliary files

show.env

logical; specify whether to print all environmental parameters

silent

logical; specify whether to search for paths for external software

Details

The function will try to locate external software tools under either Windows or Unix platform and then save the results to the plotKML.opts environment. plotKML-package does not look automatically for software paths (unless you specify this manually in your "Rprofile.site").
The external software tools are not required by default and most of operations in plotKML-package can be run without using them. GDAL, SAGA GIS and Python are highly recommended, however, for processing large data sets. The function paths looks for GDAL, ImageMagick, Python, SAGA GIS, in the Windows Registry Hive, the Program Files directory or the usr/bin installation (Unix).

References

ImageMagick (https://imagemagick.org)

GDAL (https://gdal.org)

SAGA GIS (https://saga-gis.sourceforge.io/)

Python (https://www.python.org)

Author

Tomislav Hengl, Dylan Beaudette

Note

To further customize the plotKML options, consider putting:
library(plotKML); plotKML.env(..., show.env = FALSE)
in your "/etc/Rprofile.site".

Warning

Under Linux OS you need to install GDAL binaries by using e.g.:
sudo apt-get install gdal-bin

Examples

if (FALSE) ## look for paths:
pts <- paths()
pts
#> Error in eval(expr, envir, enclos): object 'pts' not found
plotKML.env(silent = FALSE)
#> $colour_scale_numeric
#> [1] "#2C7BB6" "#ABD9E9" "#FFFFBF" "#FDAE61" "#D7191C"
#> 
#> $colour_scale_factor
#> [1] "#E41A1C" "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00" "#FFFF33" "#A65628"
#> [8] "#F781BF" "#999999"
#> 
#> $colour_scale_svar
#> [1] "#FEEDDE" "#FDBE85" "#FD8D3C" "#E6550D" "#A63603"
#> 
#> $ref_CRS
#> [1] "+proj=longlat +datum=WGS84 +no_defs"
#> 
#> $NAflag
#> [1] -99999
#> 
#> $icon
#> [1] "icon3.png"
#> 
#> $LabelScale
#> [1] 0.5
#> 
#> $size_range
#> [1] 0.25 2.50
#> 
#> $license_url
#> [1] "http://creativecommons.org/licenses/by/4.0/"
#> 
#> $metadata_sel
#>  [1] "idinfo.citation.citeinfo.title"                                       
#>  [2] "idinfo.descript.abstract"                                             
#>  [3] "spdoinfo.ptvctinf.sdtsterm.ptvctcnt"                                  
#>  [4] "idinfo.timeperd.timeinfo.rngdates.begdate"                            
#>  [5] "idinfo.timeperd.timeinfo.rngdates.enddate"                            
#>  [6] "distinfo.stdorder.digform.digtopt.onlinopt.computer.networka.networkr"
#>  [7] "idinfo.citation.citeinfo.othercit"                                    
#>  [8] "idinfo.citation.citeinfo.onlink"                                      
#>  [9] "idinfo.datacred"                                                      
#> [10] "distinfo.distrib.cntinfo.cntorgp.cntorg"                              
#> [11] "distinfo.stdorder.digform.digtinfo.formcont"                          
#> [12] "idinfo.native"                                                        
#> 
#> $kmz
#> [1] FALSE
#> 
#> $kml_xsd
#> [1] "https://schemas.opengis.net/kml/2.2.0/ogckml22.xsd"
#> 
#> $kml_url
#> [1] "https://www.opengis.net/kml/2.2/"
#> 
#> $kml_gx
#> [1] "https://www.google.com/kml/ext/2.2"
#> 
#> $gpx_xsd
#> [1] "https://www.topografix.com/GPX/1/1/gpx.xsd"
#> 
#> $fgdc_xsd
#> [1] "https://fgdcxml.sourceforge.net/schema/fgdc-std-012-2002/fgdc-std-012-2002.xsd"
#> 
#> $inspire_xsd
#> [1] "https://inspire.ec.europa.eu/schemas/common/1.0/common.xsd"
#> 
#> $convert
#> [1] "\"C:\\PROGRA~1\\IMAGEM~1.8-Q\\convert.exe\""
#> 
#> $gdalwarp
#> [1] "\"C:\\PROGRA~1\\QGIS32~1.3\\bin\\gdalwarp.exe\""
#> 
#> $gdal_translate
#> [1] "\"C:\\PROGRA~1\\QGIS32~1.3\\bin\\GDAL_T~1.EXE\""
#> 
#> $python
#> [1] ""
#> 
#> $home_url
#> [1] "https://plotkml.r-forge.r-project.org/"
#> 
gdalwarp <- get("gdalwarp", envir = plotKML.opts)
## if missing you need to install it!
system(paste(gdalwarp, "--help-general")) 
#> [1] 0
system(paste(gdalwarp, "--formats"), intern = TRUE)
#>   [1] "Supported Formats:"                                                                 
#>   [2] "  ECW -raster- (rov): ERDAS Compressed Wavelets (SDK 5.5)"                          
#>   [3] "  JP2ECW -raster,vector- (rov): ERDAS JPEG2000 (SDK 5.5)"                           
#>   [4] "  GeoRaster -raster- (rw+s): Oracle Spatial GeoRaster"                              
#>   [5] "  HDF5 -raster,multidimensional raster- (rovs): Hierarchical Data Format Release 5" 
#>   [6] "  HDF5Image -raster- (rov): HDF5 Dataset"                                           
#>   [7] "  BAG -raster,multidimensional raster,vector- (rw+v): Bathymetry Attributed Grid"   
#>   [8] "  MG4Lidar -raster- (ro): MrSID Generation 4 / Lidar (.sid)"                        
#>   [9] "  MrSID -raster- (rov): Multi-resolution Seamless Image Database (MrSID)"           
#>  [10] "  JP2MrSID -raster- (rov): MrSID JPEG2000"                                          
#>  [11] "  VRT -raster,multidimensional raster- (rw+v): Virtual Raster"                      
#>  [12] "  DERIVED -raster- (ro): Derived datasets using VRT pixel functions"                
#>  [13] "  GTiff -raster- (rw+vs): GeoTIFF"                                                  
#>  [14] "  COG -raster- (wv): Cloud optimized GeoTIFF generator"                             
#>  [15] "  NITF -raster- (rw+vs): National Imagery Transmission Format"                      
#>  [16] "  RPFTOC -raster- (rovs): Raster Product Format TOC format"                         
#>  [17] "  ECRGTOC -raster- (rovs): ECRG TOC format"                                         
#>  [18] "  HFA -raster- (rw+v): Erdas Imagine Images (.img)"                                 
#>  [19] "  SAR_CEOS -raster- (rov): CEOS SAR Image"                                          
#>  [20] "  CEOS -raster- (rov): CEOS Image"                                                  
#>  [21] "  JAXAPALSAR -raster- (rov): JAXA PALSAR Product Reader (Level 1.1/1.5)"            
#>  [22] "  GFF -raster- (rov): Ground-based SAR Applications Testbed File Format (.gff)"     
#>  [23] "  ELAS -raster- (rw+v): ELAS"                                                       
#>  [24] "  ESRIC -raster- (rov): Esri Compact Cache"                                         
#>  [25] "  AIG -raster- (rov): Arc/Info Binary Grid"                                         
#>  [26] "  AAIGrid -raster- (rwv): Arc/Info ASCII Grid"                                      
#>  [27] "  GRASSASCIIGrid -raster- (rov): GRASS ASCII Grid"                                  
#>  [28] "  ISG -raster- (rov): International Service for the Geoid"                          
#>  [29] "  SDTS -raster- (rov): SDTS Raster"                                                 
#>  [30] "  DTED -raster- (rwv): DTED Elevation Raster"                                       
#>  [31] "  PNG -raster- (rwv): Portable Network Graphics"                                    
#>  [32] "  JPEG -raster- (rwv): JPEG JFIF"                                                   
#>  [33] "  MEM -raster,multidimensional raster- (rw+): In Memory Raster"                     
#>  [34] "  JDEM -raster- (rov): Japanese DEM (.mem)"                                         
#>  [35] "  GIF -raster- (rwv): Graphics Interchange Format (.gif)"                           
#>  [36] "  BIGGIF -raster- (rov): Graphics Interchange Format (.gif)"                        
#>  [37] "  ESAT -raster- (rov): Envisat Image Format"                                        
#>  [38] "  BSB -raster- (rov): Maptech BSB Nautical Charts"                                  
#>  [39] "  XPM -raster- (rwv): X11 PixMap Format"                                            
#>  [40] "  BMP -raster- (rw+v): MS Windows Device Independent Bitmap"                        
#>  [41] "  DIMAP -raster- (rovs): SPOT DIMAP"                                                
#>  [42] "  AirSAR -raster- (rov): AirSAR Polarimetric Image"                                 
#>  [43] "  RS2 -raster- (rovs): RadarSat 2 XML Product"                                      
#>  [44] "  SAFE -raster- (rov): Sentinel-1 SAR SAFE Product"                                 
#>  [45] "  PCIDSK -raster,vector- (rw+v): PCIDSK Database File"                              
#>  [46] "  PCRaster -raster- (rw+): PCRaster Raster File"                                    
#>  [47] "  ILWIS -raster- (rw+v): ILWIS Raster Map"                                          
#>  [48] "  SGI -raster- (rw+v): SGI Image File Format 1.0"                                   
#>  [49] "  SRTMHGT -raster- (rwv): SRTMHGT File Format"                                      
#>  [50] "  Leveller -raster- (rw+v): Leveller heightfield"                                   
#>  [51] "  Terragen -raster- (rw+v): Terragen heightfield"                                   
#>  [52] "  GMT -raster- (rw): GMT NetCDF Grid Format"                                        
#>  [53] "  netCDF -raster,multidimensional raster,vector- (rw+s): Network Common Data Format"
#>  [54] "  HDF4 -raster,multidimensional raster- (ros): Hierarchical Data Format Release 4"  
#>  [55] "  HDF4Image -raster- (rw+): HDF4 Dataset"                                           
#>  [56] "  ISIS3 -raster- (rw+v): USGS Astrogeology ISIS cube (Version 3)"                   
#>  [57] "  ISIS2 -raster- (rw+v): USGS Astrogeology ISIS cube (Version 2)"                   
#>  [58] "  PDS -raster- (rov): NASA Planetary Data System"                                   
#>  [59] "  PDS4 -raster,vector- (rw+vs): NASA Planetary Data System 4"                       
#>  [60] "  VICAR -raster,vector- (rw+v): MIPL VICAR file"                                    
#>  [61] "  TIL -raster- (rov): EarthWatch .TIL"                                              
#>  [62] "  ERS -raster- (rw+v): ERMapper .ers Labelled"                                      
#>  [63] "  JP2OpenJPEG -raster,vector- (rwv): JPEG-2000 driver based on OpenJPEG library"    
#>  [64] "  L1B -raster- (rovs): NOAA Polar Orbiter Level 1b Data Set"                        
#>  [65] "  FIT -raster- (rwv): FIT Image"                                                    
#>  [66] "  GRIB -raster,multidimensional raster- (rwv): GRIdded Binary (.grb, .grb2)"        
#>  [67] "  RMF -raster- (rw+v): Raster Matrix Format"                                        
#>  [68] "  WCS -raster- (rovs): OGC Web Coverage Service"                                    
#>  [69] "  WMS -raster- (rwvs): OGC Web Map Service"                                         
#>  [70] "  MSGN -raster- (rov): EUMETSAT Archive native (.nat)"                              
#>  [71] "  RST -raster- (rw+v): Idrisi Raster A.1"                                           
#>  [72] "  INGR -raster- (rw+v): Intergraph Raster"                                          
#>  [73] "  GSAG -raster- (rwv): Golden Software ASCII Grid (.grd)"                           
#>  [74] "  GSBG -raster- (rw+v): Golden Software Binary Grid (.grd)"                         
#>  [75] "  GS7BG -raster- (rw+v): Golden Software 7 Binary Grid (.grd)"                      
#>  [76] "  COSAR -raster- (rov): COSAR Annotated Binary Matrix (TerraSAR-X)"                 
#>  [77] "  TSX -raster- (rov): TerraSAR-X Product"                                           
#>  [78] "  COASP -raster- (ro): DRDC COASP SAR Processor Raster"                             
#>  [79] "  R -raster- (rwv): R Object Data Store"                                            
#>  [80] "  MAP -raster- (rov): OziExplorer .MAP"                                             
#>  [81] "  KMLSUPEROVERLAY -raster- (rwv): Kml Super Overlay"                                
#>  [82] "  WEBP -raster- (rwv): WEBP"                                                        
#>  [83] "  PDF -raster,vector- (rw+vs): Geospatial PDF"                                      
#>  [84] "  Rasterlite -raster- (rwvs): Rasterlite"                                           
#>  [85] "  MBTiles -raster,vector- (rw+v): MBTiles"                                          
#>  [86] "  PLMOSAIC -raster- (ro): Planet Labs Mosaics API"                                  
#>  [87] "  CALS -raster- (rwv): CALS (Type 1)"                                               
#>  [88] "  WMTS -raster- (rwv): OGC Web Map Tile Service"                                    
#>  [89] "  SENTINEL2 -raster- (rovs): Sentinel 2"                                            
#>  [90] "  MRF -raster- (rw+v): Meta Raster Format"                                          
#>  [91] "  PNM -raster- (rw+v): Portable Pixmap Format (netpbm)"                             
#>  [92] "  DOQ1 -raster- (rov): USGS DOQ (Old Style)"                                        
#>  [93] "  DOQ2 -raster- (rov): USGS DOQ (New Style)"                                        
#>  [94] "  PAux -raster- (rw+v): PCI .aux Labelled"                                          
#>  [95] "  MFF -raster- (rw+v): Vexcel MFF Raster"                                           
#>  [96] "  MFF2 -raster- (rw+): Vexcel MFF2 (HKV) Raster"                                    
#>  [97] "  FujiBAS -raster- (rov): Fuji BAS Scanner Image"                                   
#>  [98] "  GSC -raster- (rov): GSC Geogrid"                                                  
#>  [99] "  FAST -raster- (rov): EOSAT FAST Format"                                           
#> [100] "  BT -raster- (rw+v): VTP .bt (Binary Terrain) 1.3 Format"                          
#> [101] "  LAN -raster- (rw+v): Erdas .LAN/.GIS"                                             
#> [102] "  CPG -raster- (rov): Convair PolGASP"                                              
#> [103] "  IDA -raster- (rw+v): Image Data and Analysis"                                     
#> [104] "  NDF -raster- (rov): NLAPS Data Format"                                            
#> [105] "  EIR -raster- (rov): Erdas Imagine Raw"                                            
#> [106] "  DIPEx -raster- (rov): DIPEx"                                                      
#> [107] "  LCP -raster- (rwv): FARSITE v.4 Landscape File (.lcp)"                            
#> [108] "  GTX -raster- (rw+v): NOAA Vertical Datum .GTX"                                    
#> [109] "  LOSLAS -raster- (rov): NADCON .los/.las Datum Grid Shift"                         
#> [110] "  NTv2 -raster- (rw+vs): NTv2 Datum Grid Shift"                                     
#> [111] "  CTable2 -raster- (rw+v): CTable2 Datum Grid Shift"                                
#> [112] "  ACE2 -raster- (rov): ACE2"                                                        
#> [113] "  SNODAS -raster- (rov): Snow Data Assimilation System"                             
#> [114] "  KRO -raster- (rw+v): KOLOR Raw"                                                   
#> [115] "  ROI_PAC -raster- (rw+v): ROI_PAC raster"                                          
#> [116] "  RRASTER -raster- (rw+v): R Raster"                                                
#> [117] "  BYN -raster- (rw+v): Natural Resources Canada's Geoid"                            
#> [118] "  ARG -raster- (rwv): Azavea Raster Grid format"                                    
#> [119] "  RIK -raster- (rov): Swedish Grid RIK (.rik)"                                      
#> [120] "  USGSDEM -raster- (rwv): USGS Optional ASCII DEM (and CDED)"                       
#> [121] "  GXF -raster- (rov): GeoSoft Grid Exchange Format"                                 
#> [122] "  NWT_GRD -raster- (rw+v): Northwood Numeric Grid Format .grd/.tab"                 
#> [123] "  NWT_GRC -raster- (rov): Northwood Classified Grid Format .grc/.tab"               
#> [124] "  ADRG -raster- (rw+vs): ARC Digitized Raster Graphics"                             
#> [125] "  SRP -raster- (rovs): Standard Raster Product (ASRP/USRP)"                         
#> [126] "  BLX -raster- (rwv): Magellan topo (.blx)"                                         
#> [127] "  PostGISRaster -raster- (rws): PostGIS Raster driver"                              
#> [128] "  SAGA -raster- (rw+v): SAGA GIS Binary Grid (.sdat, .sg-grd-z)"                    
#> [129] "  XYZ -raster- (rwv): ASCII Gridded XYZ"                                            
#> [130] "  HF2 -raster- (rwv): HF2/HFZ heightfield raster"                                   
#> [131] "  OZI -raster- (rov): OziExplorer Image File"                                       
#> [132] "  CTG -raster- (rov): USGS LULC Composite Theme Grid"                               
#> [133] "  ZMap -raster- (rwv): ZMap Plus Grid"                                              
#> [134] "  NGSGEOID -raster- (rov): NOAA NGS Geoid Height Grids"                             
#> [135] "  IRIS -raster- (rov): IRIS data (.PPI, .CAPPi etc)"                                
#> [136] "  PRF -raster- (rov): Racurs PHOTOMOD PRF"                                          
#> [137] "  RDA -raster- (ro): DigitalGlobe Raster Data Access driver"                        
#> [138] "  EEDAI -raster- (ros): Earth Engine Data API Image"                                
#> [139] "  DAAS -raster- (ro): Airbus DS Intelligence Data As A Service driver"              
#> [140] "  SIGDEM -raster- (rwv): Scaled Integer Gridded DEM .sigdem"                        
#> [141] "  TGA -raster- (rov): TGA/TARGA Image File Format"                                  
#> [142] "  OGCAPI -raster,vector- (rov): OGCAPI"                                             
#> [143] "  STACTA -raster- (rovs): Spatio-Temporal Asset Catalog Tiled Assets"               
#> [144] "  STACIT -raster- (rovs): Spatio-Temporal Asset Catalog Items"                      
#> [145] "  DB2ODBC -raster,vector- (rw+): IBM DB2 Spatial Database"                          
#> [146] "  GPKG -raster,vector- (rw+vs): GeoPackage"                                         
#> [147] "  CAD -raster,vector- (rovs): AutoCAD Driver"                                       
#> [148] "  PLSCENES -raster,vector- (ro): Planet Labs Scenes API"                            
#> [149] "  NGW -raster,vector- (rw+s): NextGIS Web"                                          
#> [150] "  GenBin -raster- (rov): Generic Binary (.hdr Labelled)"                            
#> [151] "  ENVI -raster- (rw+v): ENVI .hdr Labelled"                                         
#> [152] "  EHdr -raster- (rw+v): ESRI .hdr Labelled"                                         
#> [153] "  ISCE -raster- (rw+v): ISCE raster"                                                
#> [154] "  Zarr -raster,multidimensional raster- (rw+vs): Zarr"                              
#> [155] "  HTTP -raster,vector- (ro): HTTP Fetching Wrapper"                                 

plotKML.env(show.env = FALSE)
get("home_url", envir = plotKML.opts)
#> [1] "https://plotkml.r-forge.r-project.org/"