plotKML.env {plotKML}	R Documentation

plotKML specific environmental variables / paths

Description

Sets the environmental, package specific parameters and settings (URLs, names, default color palettes and similar) that can be later on passed to other functions.

Usage

plotKML.env(colour_scale_numeric = "", colour_scale_factor = "",
              colour_scale_svar = "", ref_CRS, NAflag, icon, LabelScale, 
              license_url, metadata_sel, kmz, kml_xsd, kml_url, kml_gx, gpx_xsd,
              fgdc_xsd, convert, gdalwarp, gdal_translate, python,
              home_url, googleAPIkey, show.env = TRUE, silent = TRUE)

Arguments

colour_scale_numeric
default colour scheme for numeric variables

colour_scale_factor
default colour scheme for factor variables

colour_scale_svar
default colour scheme for model error (e.g. mapping error)

ref_CRS
the referent CRS proj4string ("+proj=longlat +datum=WGS84")

NAflag
the default missing value flag (usually "-99999")

icon
the default icon URL

LabelScale
the default scale factor for labels

license_url
the default license URL

metadata_sel
a list of the default metadata fields for summary

kmz
logical; the default compression setting

kml_xsd
the default KML scheme URL

kml_url
the default KML format URL

kml_gx
the default extended KML scheme URL

gpx_xsd
the default GPX scheme URL

fgdc_xsd
the default metadata scheme URL

convert
a path to ImageMagick convert program

gdalwarp
a path to FWTools gdalwarp program

gdal_translate
a path to FWTools gdalwarp program

python
a path to Python program

home_url
the default location of all icons and auxiliary files

googleAPIkey
(optional) Google API key required for some operations

show.env
logical; specify whether to print all environmental parameters

silent
logical; specify whether to search for paths for external software

Details

The function will try to locate external software tools under either Windows or Unix platform and then save the results to the plotKML.opts environment. plotKML-package does not look automatically for software paths (unless you specify this manually in your "Rprofile.site").
The external software tools are not required by default and most of operations in plotKML-package can be run without using them. FWTools, SAGA GIS and Python are highly recommended, however, for processing large data sets. The function paths looks for FWTools, ImageMagick, Python, SAGA GIS, in the Windows Registry Hive, the Program Files directory or the usr/bin installation (Unix).

Warning

Under Linux OS make sure you add path to executable program by using e.g.:
export PATH=INSTALL_DIRECTORY/FWTools-2.*.*/bin_safe:$PATH

Note

To further customize the plotKML options, consider putting:
library(plotKML); plotKML.env(..., show.env = FALSE)
in your "/etc/Rprofile.site".

Author(s)

Tomislav Hengl, Dylan Beaudette

References

ImageMagick (http://imagemagick.org)

FWTools (http://fwtools.maptools.org)

SAGA GIS (http://www.saga-gis.org)

Python (http://python.org)

Examples


## Not run: # look for paths:
pts <- paths()
pts
plotKML.env(silent = FALSE)
gdalwarp <- get("gdalwarp", envir = plotKML.opts)
# if missing you need to install it!
system(paste(gdalwarp, "--help-general")) 
gdal.formats <- system(paste(gdalwarp, "--formats"), intern = TRUE)
## End(Not run)
plotKML.env(show.env = FALSE)
get("home_url", envir = plotKML.opts)

[Package plotKML version 0.2-3 Index]