Import GPX (GPS track) files

Reads various elements from a *.gpx file --- metadata, waypoints, tracks and routes --- and converts them to dataframes.

readGPX(gpx.file, metadata = TRUE, bounds = TRUE, 
              waypoints = TRUE, tracks = TRUE, routes = TRUE)

Arguments

gpx.file

location of the gpx.file

metadata

logical; species whether the metadata should be imported

bounds

logical; species whether the bounding box coordinates should be imported

waypoints

logical; species whether all waypoints should be imported

tracks

logical; species whether all tracks should be imported

routes

logical; species whether all routes should be imported

Details

Waypoint is a point of interest, or named feature on a map. Track is an ordered list of points describing a path. Route is an ordered list of waypoints representing a series of turn points leading to a destination.

References

GPX data format (https://www.topografix.com/gpx.asp)

XML tutorial (https://github.com/omegahat/XML)

Author

Tomislav Hengl

See also

Examples

if (FALSE) # read GPX file from web:
fells_loop <- readGPX("https://www.topografix.com/fells_loop.gpx")
str(fells_loop)
#> Error in str(fells_loop): object 'fells_loop' not found