Imports GBIF cell density records

Read GBIF cell (1--degree) density record counts and converts them to a "raster" object.

readKML.GBIFdensity(kml.file, gbif.url = FALSE, silent = FALSE)

Arguments

kml.file

GBIF cell density file (local file or URL)

gbif.url

logical; species whether the cellid and taxon content information should be also imported (usually not used)

silent

logical; species whether the progress bar should be printed

Details

This document contains data shared through the GBIF Network --- see https://www.gbif.org/occurrence for more information. GBIF records are constantly updated and every map derived refers to a certain date indicated in the @zname Last update slot.
All usage of these data must be in accordance with the GBIF Data Use Agreement: https://www.gbif.org/terms.

References

GBIF cell density description (https://www.gbif.org/occurrence)

Author

Tomislav Hengl

See also

Examples

if (FALSE) # reading taxon density maps:
kml.file <- "taxon-celldensity-2294100.kml"
# download.file(paste("http://data.gbif.org/occurrences/taxon/celldensity/", kml.file, sep=""),
# destfile=paste(getwd(), kml.file, sep="")) 
# this will not run (you must first accept the data usage agreeent); 
# instead, obtain the kml file via a web browser, and save it to the working directory:
r <- readKML.GBIFdensity(kml.file)
#> Error in file.info(kml.file): object 'kml.file' not found
class(r)
#> Error in eval(expr, envir, enclos): object 'r' not found
summary(r)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'summary': object 'r' not found
image(r)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'image': object 'r' not found
# add world borders:
library(maps)
#> Error in library(maps): there is no package called 'maps'
country.m = map('world', plot=FALSE, fill=TRUE)
#> Error in map("world", plot = FALSE, fill = TRUE): could not find function "map"
IDs <- sapply(strsplit(country.m$names, ":"), function(x) x[1])
#> Error in strsplit(country.m$names, ":"): object 'country.m' not found
library(maptools)
#> Checking rgeos availability: TRUE
#> Please note that 'maptools' will be retired by the end of 2023,
#> plan transition at your earliest convenience;
#> some functionality will be moved to 'sp'.
country <- as(map2SpatialPolygons(country.m, IDs=IDs), "SpatialLines")
#> Error in map2SpatialPolygons(country.m, IDs = IDs): package maps required
lines(country)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'lines': object 'country' not found
# to import a list of files, use e.g.:
kml.list <- list(kml.file)
#> Error in eval(expr, envir, enclos): object 'kml.file' not found
r.lst <- lapply(kml.list, readKML.GBIFdensity, silent = TRUE)
#> Error in lapply(kml.list, readKML.GBIFdensity, silent = TRUE): object 'kml.list' not found
# mask out missing layers (empty KML files):
mask <- !sapply(r.lst, is.null)
#> Error in lapply(X = X, FUN = FUN, ...): object 'r.lst' not found
r.lst <- brick(r.lst[mask])
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'brick': object 'r.lst' not found